==> 2_analyze_data.Rmd <==
---
header-includes:
    - \usepackage{setspace}\singlespacing
    - \usepackage[bottom]{footmisc}
    - \usepackage{float}
    - \usepackage{booktabs}
    #- \usepackage{float}\floatsetup[figure]{capposition=top}
geometry: margin=1in
output: 
  pdf_document: 
    latex_engine: xelatex
    keep_tex: yes
    number_sections: true
    fig_caption: yes
bibliography: references.bib
csl: apa-5th-edition.csl
fontsize: 12pt
linestretch: 1.15
---
\pagenumbering{gobble}
\definecolor{azulUC3M}{RGB}{0,0,102}

\begin{titlepage}
	\begin{sffamily}
	\color{azulUC3M}
	\begin{center}
		\begin{figure}[H] 
			\makebox[\textwidth][c]{\includegraphics[width=16cm]{imagenes/Portada_Logo.png}}
		\end{figure}
		\vspace{1cm}
		\begin{Large}
			Master Degree in Economic Development and Growth\\			
			2021 - 2022\\
			\vspace{1cm}		
			\textsl{Master Thesis}
			\bigskip
			
		\end{Large}
		 	{\Huge Co-ethnic density and economic outcomes:}\\
		 	{\Huge A microeconometric analysis of how residential settlement patterns impact immigrant wages using the Spanish National Survey of Immigrants}\\
		 	\vspace*{0.5cm}
	 		\rule{10.5cm}{0.1mm}\\
			\vspace*{0.9cm}
			{\LARGE Madeline Swarr}\\ 
			\vspace*{1cm}
		\begin{Large}
			Prof. Raquel Carrasco Perea\\
		\end{Large}
	\end{center}
	
	\vfill
	\color{black}
	\begin{footnotesize}
	\noindent\fbox{
	\begin{minipage}{\textwidth}
		\textbf{AVOID PLAGIARISM}\\
		The University uses the \textbf{Turnitin Feedback Studio} program within the Aula Global for the delivery of student work. This program compares the originality of the work delivered by each student with millions of electronic resources and detects those parts of the text that are copied and pasted. Plagiarizing in a TFM is considered a \textbf{Serious Misconduct}, and may result in permanent expulsion from the University.
	\end{minipage}	
	}
	\vspace*{.5cm}\\	
	\noindent\includegraphics[width=4.2cm]{imagenes/creativecommons.png}\\
	This work is licensed under Creative Commons \textbf{Attribution – Non Commercial – Non Derivatives}

	\end{footnotesize}
	\end{sffamily}
\end{titlepage}

\newpage
  
```{css, echo=FALSE}
h1 {
  text-align: center;
}
```

```{r setup, echo = FALSE, warning = FALSE, message = FALSE}
######################## Specify head directory ################
head_dir <- "/Users/papapeach/Documents/UC3M/Thesis/Data Analysis/ENI-Analysis"
run_date <- "30aug2022"

######################## (Down)load libraries ######################
dependency_dir <- paste0(head_dir, "/dependencies")
data_dir <- paste0(head_dir, "/data/")
## Set directory where processed ENI data can be found
input_dir <- paste0(head_dir, "/output/", run_date, "/1_processed")
output_dir <- paste0(head_dir, "/output/", run_date, "/2_analyzed")
# Create output directory
if (!dir.exists(output_dir)){
  dir.create(output_dir, recursive = T)
}
source(paste0(dependency_dir, "/get_packages.R"))
get_packages(c("RColorBrewer", "openxlsx", "sf", "terra", "spData", 
               "spDataLarge", "mapSpain", "tmap", "reshape", "gifski", "RColorBrewer", "kableExtra",
               "MASS", "foreign", "Hmisc", "reshape2", "stargazer", "robustbase", "hypr",
               "sjPlot", "sjmisc", "estimatr", "AER", "sampleSelection", "plyr", "tidyverse", "ggplot2", "sampleSelection",
               "extrafont"))

# Import fonts for figures and tables
#import_fonts()
loadfonts(device="win")
```

\newpage
\vspace{-5truemm}
\begin{abstract}
\noindent Social science has always been concerned with understanding the meso-level forces that connect individuals to the whole. In the case of the foreign immigrant, this integration into a new society, often characterized by differences in institutions, language, culture, and socioeconomic structure, is a long and arduous process that those of the native population do not always face, and if they do, do not face as intensely. 

\vspace{2truemm}

\noindent This paper defends the idea that pre-established migratory networks affect the residential settlement patterns of immigrants in regions where others from their home country tend to settle. While these networks are intuitively helpful in providing resources and protecting against the psychological shocks of such displacement, they can also help explain why those living in provinces with many other co-ethnics experience differential economic outcomes than those counterparts living in less concentrated regions. This analysis uses the 2007 National Survey of Immigrants and municipal registry data from Spain to empirically assess the relationship between residential co-ethnic density and wage differentials of immigrants who arrived to Spain between 1998 and 2007. I find evidence that this relationship varies depending on characteristics of the migratory network. 
\end{abstract}

\vspace{-8truemm}
\tableofcontents

\newpage
\pagenumbering{arabic}
# INTRODUCTION {#introduction}

Simon Kuznets observed in his early studies on the economic lives of migrants that “all minorities are characterized, at any given time, by an occupational structure distinctly narrower than that of the total population and the majority” [@kuznets1960economic, p. 14]. The concept of “ethnic penalty” in immigration studies describes the fact that a meritocratic society is far from reality for these populations, and that the degree of penalty on wages and labor mobility depends on the ethnic group being referenced [@hasmath2012ethnic]. Therefore, the classic determinants of economic well-being for immigrants do not have the same effects as for those of the native population, and these effects may differ between immigrant groups. 

\vspace{3truemm}

Many empirical studies have shown that immigrants experience lower returns to human capital than natives do, especially for education or experience attained in origin countries that are much different than host countries in terms of economic development, labor market structure, language, and institutions (see @friedberg2000you, @kanas2009impact, and @lequang2022returns for examples). Therefore, the occupational structure is most narrow for immigrants in the years directly following arrival in the destination country when little opportunity has been given to accumulate and assimilate the human capital necessary to participate fully in the local labor market. Given the restrictions placed on newly-arrived immigrants, the impact that migrant networks can have on integration, especially in the first few years, becomes that much greater. Even if social ties are not leveraged explicitly for personal gain, the indirect influence that networks can have on protection against discrimination, information sharing, and trust can indirectly affect one's rate of skill acquisition, job prospects, salary, and emotional well-being, as well as play a critical role in the initial decision of whether to migrate or not [@chiswick2001determinants] [@kerr2015social]. While migratory networks undoubtedly provide access to resources that ease the transition into a new country, it is less clear of whether the geographical concentration of immigrant groups by ethnicity that these networks bring about has an impact on the economic lives of immigrants, and if so, whether this impact is positive or negative and to what extent.

While much of the debate surrounding economic assimilation has centered around measuring the rate of convergence between native and immigrant wages over time, or economic *assimilation*, more recent research has focused on understanding what leads to differences in economic *well-being* between cohorts of immigrants. This paper not only investigates the effect of co-ethnic residential choices on hourly wages of recent immigrants to Spain, but it also sheds light on how migrant networks, actors in the settlement decision process, may be determining this effect. Utilizing detailed micro data from the 2007 National Survey of Immigrants (*Encuesta nacional de inmigrantes*) and linking them to population data from the municipal registry database (*Estadística del Padrón continuo*) to draw correlations between immigrant wages and residential co-ethnic density, I find the effect of living in a province with a higher percentage of co-ethnics to be consistently negative and robust after controlling for certain individual-level characteristics and other fixed effects such as location and country of origin. However, I also find evidence that relative concentration of co-ethnics has the opposite effect. These conclusions remain unchanged even after addressing endogeneity issues that arise when studying voluntary migration.

The rest of the paper is structured as follows: the remainder of Section \ref{introduction} discusses the theoretical framework of the study and contextualizes the debate within Spain. Section \ref{data} describes the data and study parameters used for the empirical analysis. Section \ref{methodology} outlines the wage estimation methodology. Section \ref{results} presents the results for both the primary and secondary analyses. Finally, the paper concludes with a general remarks in Section \ref{conclusion}.

## Theoretical Framework {#theoretical_framework}
Many of the neoclassical models of migration predominant in the literature over the last century stem from the framework laid out by British geographer Ernst Ravenstein in his 1885 address to the Royal Statistical Society of London, *The Laws of Migration*. What may be better referred to as a set of eleven hypotheses as to why people do or do not migrate has been synthesized over time into a set of "push" and "pull" factors, the former representing those conditions in one's present location that make life difficult or unappealing, and the latter representing those conditions in a potential new location that are attractive or more desirable in comparison [@dorigo1983push]. Accepting the assumption that humans are utility maximizers, the migratory decision is thus understood as an economic one, a simple cost-benefit analysis to see if the pros of emigrating versus staying outweigh the cons. In support of this idea, much economic theory and empirical studies have shown that wage and employment condition differentials are key determinants in predicting the unforced flow of migrants between specific origin-destination dyads [@harris1970migration] [@rabe2012differences]. However, if these were the only determinants, we would expect to see a much greater proportion of the world's population living abroad than the mere 3% we see today given the current economic divide between developed and developing countries [@wmr2020]. 

Why then do so many people choose to stay home rather than making the big leap? Not only can the physical and psychological costs of uprooting and replanting oneself in a foreign country be immense, but political, cultural, and informational barriers may make it impossible for one to even envision what may be waiting for them on the other side. While advances in transportation technology and changes to immigration policy also have an effect on the direction and magnitude of migratory flows, increasing emphasis in modern migration discourse has been placed on the role of co-ethnic migratory networks on the migratory process. Chain migration, which @macdonald1964chain define as the process by which "prospective migrants learn of opportunities, are provided with transportation, and have initial accommodation and employment arranged *by means of primary social relationships with previous migrants*" help explain why settlement processes of immigrants, once arrived in their destination, result in relative concentration by ethnicity or foreign nationality in specific regions of the host country (p. 84). These ethnic enclaves, as they are often referred to, have been the subject of great debate amongst scholars and policymakers alike about whether they are beneficial or harmful to the economic integration of the immigrant. 

On one hand, the enclave can represent a "warm embrace" in what may otherwise be a rather hostile environment to a recent transplant [@borjas2000ethnic]. The importance of social capital amongst immigrants for correcting for information imbalances, resource scarcity, cultural displacement, restricted rights, and discrimination has been documented extensively in the literature (see @massey1987social, @munshi2003networks, and @epstein2010migration for examples). Geographic clustering of immigrant groups by ethnicity can provide valuable economic support through job contacts, as well as create jobs directly if groups decide to establish local economies that serve their own "ethnic market" [@portes198113, p. 291]. Other studies have shown that co-ethnic networks can breed productivity, entrepreneurship, skills acquisition, and innovation through trust and information-sharing that arises from a sense of shared identity and similar tastes (see @light1990korean, @ozgen2012immigration, and @kerr2015social for examples). On the other hand, it is also argued that ethnic clustering results in, or is a result of, segregation from and discrimination by the institutional systems of the host society at large that push these groups into the secondary sector characterized by temporary contracts, minimum wage, poor working conditions, and little opportunity for promotion (see @doeringer1975unemployment for more). Ethnic enclaves can thus act as a double-edged sword -- while providing an easy gateway into a new country, they can also hinder job mobility by restricting access to the more lucrative jobs of the primary labor market, reduce incentives to build competitive skills like local language ability, and create greater job competition and depress wages by the extent to which labor is substitutable within and free-flowing into the enclave [@borjas1994ethnicity].

Since the mechanisms are theoretically various and directionally ambiguous, researches have turned to empirics to estimate the overall net effect of co-ethnic density on economic outcomes with various findings that differ by context and situation. @portes1985latin have studied extensively the situation of Cuban immigrants in Miami, Florida and have generally found greater returns for Cuban-owned businesses and their co-ethnic employees than for their counterparts working outside of the ethnic enclave, while @sanders1987limits have argued that it is only ethnic business owners who benefit from the solidarity and insularity of the enclave. @borjas2000ethnic tracks immigrant cohorts in 1980 and 1990 US Census data and find that groups who settled in metropolitan regions with higher concentration of co-ethnics experienced lower wage growth over time in comparison with their counterparts who settled in less concentrated regions. 

This paper contributes to the discussion above by attempting to isolate the independent effect of residential co-ethnic concentration on wages for recent immigrants to Spain. While initial results suggest that this effect is negative and significant, closer analysis reveals that discernment between certain qualities of the co-ethnic network may be essential to further decompose this effect. 

## Contextual Motivation
While many studies on migrant concentration and economic well-being have been situated in countries with long histories as migrant destinations such as the United States, Canada, and Australia, research has increasingly focused on the Spanish context given the fact that it is now home to one of the largest immigrant populations in the world.^[As measured by percentage of the resident population.] According to the IOM’s World Migration Report, Spain was Europe's sixth most popular migrant destination in 2019 with nearly 13% of its population being made up of foreign-born individuals [@wmr2020]. Morocco, Romania, United Kingdom, and Colombia represent the top countries of origin as of the second quarter of 2021 [@inepressrelease]. However, all of these estimates vary due to the counting issues that arise from unregistered immigrants, as well as how an immigrant is defined. For example, an immigrant may be defined as anyone who is foreign-born (who may or may not have Spanish nationality), or anyone who possesses foreign citizenship. The estimates also vary widely depending on the geographical region within Spain. Nearly five hundred municipalities in 2022 (6%) reported populations with over 20% of all residents born in a different country [@inepressrelease2022]. 

Spain as net importer of migrants is a trend that picked up speed around the turn of the twentieth century as a response to the prolonged period of growth the Spanish economy was experiencing at the time. Unemployment rates decreased drastically from nearly 25% in 1994 to record-low of around 8% at the beginning of 2007. Increased foreign investments in real estate paired with an increasingly educated but thinning youth labor force pried open a huge labor demand in the construction, agriculture, and hospitality sectors that many immigrants were able to fill. This period was also characterized by increases in female labor force participation rates, a trend further fueled by the large inflow of immigrants, mainly female and originating from Andean countries, that were meeting the new demand for domestic labor as women spent less time at home and more time at work. Immigrants contributed an estimated one-third of the total national GDP between 2000 to 2007 not only through the supply of much needed low-skilled labor, but through direct increases in consumption and demand for housing as well [@galvez2020size]. 

From the 10-year period between 1998 to 2007 alone, the proportion of foreign-born persons registered as living in Spain increased over five-fold, from 1.6% to 10% of the total population. Estimates of the true percentage of foreign-born inhabitants, not just those registered in the municipal system, over the years 1970 and 2020 are shown in Figure 1 below. 

```{r echo = FALSE, warning = FALSE, message = FALSE}
pop_data <- read.xlsx(paste0(data_dir, "/series_temporal_poblacion_inmigrante.xlsx"), sheet = 1, startRow = 17, cols = c(1:4)) %>% melt(id.vars = c("Year", "Proportion")) %>% 
  dplyr::rename(Count = value, Population = variable) %>% 
  mutate(Count_1000000 = Count/1000000,
         Percent = Proportion*100)
```

```{r figpop, fig.pos = "H", echo = FALSE, warning = FALSE, message = FALSE, results='hide', fig.height = 4, fig.width = 6, fig.align = "center", fig.cap = "Counts of Spanish population by origin and percentage foreign-born (1970-2020)"}
ggplot(pop_data %>% filter(!(Year %in% c(1960, 1965)))) + geom_col(aes(x = Year, y = Count_1000000, fill = Population)) + geom_line(aes(x = Year, y = 2*Percent), size = 1.5, color = "red") + 
  scale_y_continuous(name = "Count (per million)", sec.axis = sec_axis(~./2, name = "% of Population Foreign-Born")) +
  scale_x_continuous(breaks = c(1970, 1975, 1980, 1985, 1990, 1995, 2000, 2005, 2010, 2015, 2020)) +
  theme_bw(base_size = 12) +
  labs(caption = "Data source: UN - World Population Prospects") +
  theme(plot.caption = element_text(hjust = 0),
        #text = element_text(family = "Times New Roman"),
        legend.position="bottom")

```

The rapid influx of immigrants during this time presents a unique setting to study how initial settlement patterns of individuals, often determined with little understanding of local contexts and dictated by broader forces of the migratory network, can have an impact on the economic integration of recently arrived migrants in Spain. 

\vspace{-3truemm}
# DATA {#data}
## *Encuesta nacional de inmigrantes* (ENI) from 2007
Primary data for this analysis comes from the Spanish National Institute of Statistics' (INE) National Survey of Immigrants (ENI). The survey population consists of 15,465 immigrant households living in Spain as of January 2007 that were randomly selected from predesignated pools of households with at least one individual registered as living there who met the following requirements: The individual 1) was born outside of Spain; 2) was 16 years or older at time of surveying; 3) had established residency in Spain for at least one year prior to surveying and if not the case, had the intention of staying in Spain for at least a year; and 4) for those who had Spanish nationality from birth (but born outside of Spain), must had been at least 2-years-old at the time of arrival to Spain. 

The sampling process was defined at the autonomous community level (NUTS-2) on differing population strata to target certain groups of nationalities of interest, particularly those from Ecuador, Morocco, Romania, Latin American countries excluding Ecuador, African countries excluding South Africa, Asian countries excluding Japan, North American excluding Mexico and Oceania, and the EU15 and European Economic Area excluding EU member countries. 87.37% of the desired sample size was achieved (15,465 of 17,700), with 57.62% of surveyed households coming from the primary targeted population and 42.38% coming from the supplementary targeted population. 144 countries of origin are captured in the final ENI sample, with Moroccans, Romanians, Ecuadorians, and Colombians being the most represented nationalities (12.0%, 8.6%, 8.5%, and 6.7%, respectively). All 50 Spanish provinces (NUTS-3) are represented in the sample, with Madrid, Barcelona, Illes Baleares, and Murcia being the most popular provinces of residence (12.1%, 8.5%, 7.6%, and 7.1%, respectively).

When it comes to characterizing the cohorts of immigrants arriving to Spain, the ENI is advantageous in that its highly detailed questionnaire captures over 1,500 columns of data that identify the "demographic and social characteristics of those persons born abroad, as well as their migratory itineraries, work and residential history, family relations and relations with their country of origin and with the Spanish society" [@reher2008informe].^[Link to the raw data, dictionaries, survey methodology, and questionnaire used to collect responses can be found at https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177005&menu=resultados&idp=1254735573002#!tabs-1254736195389] Missing information is, however, one downfall of such a broadly targeted and highly detailed survey. Mainly, not all who were surveyed in the ENI were employed at the time as the survey was not focused solely on measuring immigrants' economic lives. Current income and employment information is available for about half of respondents (49.7%) while employment history is available for about two-thirds of respondents (65.7%). While other annual surveys such as the Annual Wage Structure Survey and Continuous Household Survey also provide micro-level information on Spanish residents including those foreign-born, neither is designed to measure the social, economic, and demographic structure of *immigrants* exclusively, nor do they report explicit data on job-related earnings or detailed data on migratory and settlement history. In this sense, the ENI provides an incomparable level of detail that can be used to understand better the potential effects of co-location of immigrants on individual and group-level economic outcomes.^[More detailed information about the ENI's surveying methods, descriptive statistics, and diagnostics can be found in @reher2009national.]

The broad definition of "immigrant" employed in the ENI begs the need of differentiating between different types of foreign residents and their underlying motivations for coming to live in Spain. After all, there are most likely underlying differences between those who decide to locate themselves where they expect good job prospects to be versus those who, say, are looking for a sunny beach to retire to. Migratory motives are provided in the ENI data, so I choose to filter to the 8,604 individuals I deem "economic immigrants", or those who have come to Spain for 1) lack of work in home country, 2) to look for better employment, or 3) for higher quality of life so long as they were not from a country classified as "developed".^[Among the other options to choose from were for retirement, political or religious reasons, family reunification, formative or educational reasons, climate, or for a temporal stay while transitioning to another country. An individual could check off two or more motives if relevant. Therefore, to ensure that a retiree who answered as their reasons for moving to Spain as both for retirement *and* for lack of work in home country does not get labelled as an economic immigrant, only those who checked off just one or both of the two aforementioned motives qualify.] ^[Developed countries are defined as those that had a recorded GDP per capita that was at least 75% or more of that of Spain in 2007.] Making the distinction between economic and non-economic immigrants helps ensure that any unobserved factors that could simultaneously affect settlement patterns *and* economic outcomes are at least arguably more similar within these groups than across them, and therefore easier to address. To check the internal validity of this measure of motivation, I conduct a two-sample t-test and find that the proportion of those currently working at the time of surveying is significantly greater for the sample of economic immigrants than that for sample of the non-economic immigrants (Table \ref{tab:tabprops}). 

```{r echo = FALSE, warning = FALSE, message = FALSE}
if (file.exists(file = paste0(output_dir, "/datos_eni_07_processed_for_analysis.rds"))) {
  this_data <- readRDS(paste0(output_dir, "/datos_eni_07_processed_for_analysis.rds"))
  
} else {
  ## Load processed ENI data
  this_data <- readRDS(paste0(input_dir, "/datos_eni_07_processed.rds"))
  
  # Factor categorical variables and order the factors
  this_data$education <- factor(this_data$education, levels = c("Alguno/no sabe", "Sin estudios", "Primaria (incompleto)", "Primaria (completo)", "Secudaria (primer ciclo)", "Secudaria (segundo ciclo)", "Terciaria (primer ciclo)", "Terciaria (segundo ciclo)")) %>% as.factor()
  this_data$spanish <- factor(this_data$spanish, levels = c("Necesita mejorar", "Suficiente", "Bien", "Nativo")) %>% as.factor()
  this_data$sex <- factor(this_data$sex, levels = c("Male", "Female")) %>% as.factor()
  this_data$work_permit <- factor(this_data$work_permit, levels = c("Yes", "No"))
  this_data$sector2 <- factor(this_data$sector2, levels = c("Agriculture", "Business, commerce", "Construction", "Education", "Energy", "Extractive industries", "Financial intermediation", "Health, social, and community services", "Hospitality", "Household activities", "Manufacturing", "Public administration, defense", "Real estate; rental, business services", "Transportation, communications", "Unspecified")) 
  
  # Convert education to numeric factors
  this_data$educ_num <- mapvalues(this_data$education, from = c("Sin estudios", "Alguno/no sabe", "Primaria (incompleto)", "Primaria (completo)", "Secudaria (primer ciclo)", "Secudaria (segundo ciclo)", "Terciaria (primer ciclo)", "Terciaria (segundo ciclo)"), to = c(0, 1, 2, 3, 4, 5, 6, 7)) %>% as.numeric()
  # Translate education-level classification
  this_data$education_eng <- mapvalues(this_data$education, from = c("Sin estudios", "Alguno/no sabe", "Primaria (incompleto)", "Primaria (completo)", "Secudaria (primer ciclo)", "Secudaria (segundo ciclo)", "Terciaria (primer ciclo)", "Terciaria (segundo ciclo)"), to = c("No studies", "Some studies", "Primary (incomplete)", "Primary (complete)", "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)"))
  # Create hierarchical ordering of translated education levels
  this_data$education_eng <- factor(this_data$education_eng, levels = c("No studies", "Some studies", "Primary (incomplete)", "Primary (complete)", "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)"))
  # Group education-level from completed secondary education degree onward
  this_data$education_eng_class <- mapvalues(this_data$education, from = c("Sin estudios", "Alguno/no sabe", "Primaria (incompleto)", "Primaria (completo)", "Secudaria (primer ciclo)", "Secudaria (segundo ciclo)", "Terciaria (primer ciclo)", "Terciaria (segundo ciclo)"), to = c("No secondary", "No secondary", "No secondary", "No secondary", "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)"))
  # Order
  this_data$education_eng_class <- factor(this_data$education_eng_class, levels = c("No secondary", "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)"))

  # Make a new column indicating whether immigrant is an "economic immigrant"...if they indicated that they came for higher quality of living, they must have not come from a developed country
  this_data <- this_data %>% mutate(economic_final = ifelse(economic == 1 | (quality == 1 & country_class != "Developed countries (>75% GDPpc of Spain)"), TRUE, FALSE))
  this_data <- this_data %>% mutate(working = mapvalues(working, from = c(1, 0), to = c("Yes", "No")))
  
  ## IMPUTATION WEEKLY HOURS
  weekly_hours_model <- lm(weekly_hours ~ eth_dens_prov_1 + country + comunidad + education_eng_class + years + age + age_sq + sex + work_permit + sector2 , data = this_data %>% filter(economic_final == T & !is.na(weekly_hours)))
  this_data$weekly_hours_imputed[this_data$economic_final == T & is.na(this_data$weekly_hours)] <- predict(weekly_hours_model, data = this_data %>% filter(economic_final == T & is.na(weekly_hours)))
  this_data$weekly_hours_final <- ifelse(is.na(this_data$weekly_hours) & this_data$economic_final == T, this_data$weekly_hours_imputed, this_data$weekly_hours)
  this_data$hour_earn_final <- this_data$month_earn/(this_data$weekly_hours_final*4)
  this_data$log_hour_earn_final <- log(this_data$hour_earn_final)
  ## IMPUTATION AGE 
  age_model <- lm(age ~ eth_dens_prov_1 + country + comunidad + education_eng_class + years + sex + work_permit + sector2 , data = this_data %>% filter(economic_final == T & !is.na(age)))
  this_data$age_imputed[this_data$economic_final == T & is.na(this_data$age)] <- predict(age_model, data = this_data %>% filter(economic_final == T & is.na(age)))
  this_data$age_final <- ifelse(is.na(this_data$age) & this_data$economic_final == T, this_data$age_imputed, this_data$age)
  this_data$age_sq_final <- (this_data$age_final)^2
  ## IMPUTATION YEARS
  years_model <- lm(years ~ eth_dens_prov_1 + country + comunidad + education_eng_class + age + sex + work_permit + sector2 , data = this_data %>% filter(economic_final == T & !is.na(years)))
  this_data$years_imputed[this_data$economic_final == T & is.na(this_data$years)] <- predict(years_model, data = this_data %>% filter(economic_final == T & is.na(years)))
  this_data$years_final <- ifelse(is.na(this_data$years) & this_data$economic_final == T, this_data$years_imputed, this_data$years)
  this_data$years_sq_final <- (this_data$years_final)^2
  
  # Make new variable representing population stratification within province
  this_data$province_strata <- paste0(this_data$province, this_data$mun_strata)
  
  # Make a new column indicating whether row has complete cases in the columns we are interested in for
  # main wage analysis (Analysis 1)
  this_data$complete_case_1 = complete.cases(this_data %>% dplyr::select(education_eng, education_eng_class, years_final, age_final, age_sq_final, sex, country, country_class, comunidad, province, municipality))
  # Make a new column indicating whether row has complete cases in the columns we are interested in for
  # IV wage analysis (Analysis IV)
  this_data$complete_case_iv = complete.cases(this_data %>% dplyr::select(education_eng, education_eng_class, years_final, age_final, age_sq_final, sex, country, country_1991, country_class, comunidad, province, municipality))
  # Make a new column indicating whether row has complete cases in the columns we are interested in for
  # job search analysis (Analysis 2)
  this_data$complete_case_2 = complete.cases(this_data %>% dplyr::select(lt_search_3mo, education_eng, education_eng_class, educ_num, years_job_search, age_job_search, age_job_search_sq, sex, had_worked, country_class, sector1, occupation1, tasa_paro_first_job, province_job_search))

  ## Save categories and factoring
  saveRDS(this_data, file = paste0(output_dir, "/datos_eni_07_processed_for_analysis.rds"))
}

# Population of those with wage data...education, years in Spain, age, sex country of origin..Filter to those who come from countries represented at least twice in the ENI population
this_pop_1 <- this_data %>% 
  filter(economic_final == T & complete_case_1 == T & ANOM_floor >= 1998 & !(province %in% c("Ceuta", "Melilla"))) %>% 
  mutate(wage_data = ifelse(!is.na(log_hour_earn), 1, 0),
         woman_married_kids = ifelse(sex == "Female" & dependents > 0 & partner_convive == 1, 1, 0)) %>%
  group_by(country) %>% filter(n() > 10) %>% ungroup() %>%
  group_by(province_strata) %>% filter(n() > 11) %>% ungroup()

# Population for time to find a first job analysis
this_pop_2 <- this_data %>% group_by(country) %>% filter(n() > 1) %>% ungroup() %>% 
  dplyr::filter(economic_final == T & precontract == 0 & first_job_still == 1 & complete_case_2 == T & !is.na(log_eth_dens_prov_2) & !is.infinite(log_eth_dens_prov_2)) 
## Prior work experience, prior Spain experience?, Spanish ability, work status upon arrival? education level upon arrival, comunidad or province control?
```

```{r echo = FALSE, warning = FALSE, message = FALSE}
this_pop_economic <- this_data %>% dplyr::filter(economic_final == T) 

this_pop_non_economic <- this_data %>% dplyr::filter(economic_final == F)

prop_test_function <- function(pop1, pop2, var_list, alt_hyp = "two.sided") {
  results = c()
  for (var in var_list) {
    this_var_pop1 = pop1 %>% dplyr::select(var) %>% filter(!is.na(.)) %>% pull(var)
    this_var_pop2 = pop2 %>% dplyr::select(var) %>% filter(!is.na(.)) %>% pull(var)
    denom1 = length(this_var_pop1)
    denom2 = length(this_var_pop2)
     these_denoms = c(denom1, denom2)
    for (level in levels(factor(this_var_pop1))){
      numer1 = sum(factor(this_var_pop1) == level)
      numer2 = sum(factor(this_var_pop2) == level)
      these_numers = c(numer1, numer2)
      test = prop.test(x = these_numers, n = these_denoms, p = NULL, alternative = alt_hyp, correct = TRUE)
      results = rbind(results, c(var, level, numer1, denom1, round(test$estimate[1]*100, 2), numer2, denom2,
                                 round(test$estimate[2]*100, 2), signif(test$p.value, 3)))
    }
  }
  return(results %>% as.data.frame() %>% dplyr::rename(Var1 = V1, var = V2, count_econ = V3, economic = V4, percent_econ = 'prop 1', count_nonecon = V6, noneconomic = V7, percent_nonecon = 'prop 2', p = V9))
}

prop_test_results <- prop_test_function(this_pop_economic, this_pop_non_economic, var_list = c("working", "sex", "education_eng_class", "country_class", "sector2", "comunidad")) %>%
  filter(!(var %in% c("Education", "Energy", "Extractive industries", "Financial intermediation", "Public administration, defense", "Transportation, communications"))) %>%
  dplyr::select(-Var1)
  #mutate(Var1 = mapvalues(Var1, from = c("working", "sex", "education_eng_class", "country_class", "sector2", "comunidad"), to = c("Currently working?", "Sex", "Level of studies", "Country classification", "Top sectors of work", "Autonomous Community"))) 

```

```{r tabprops, echo = FALSE, warning = FALSE, message = FALSE, fig.align = "center"}
kable(prop_test_results, format = "latex", caption = "Comparison of Economic and Non-Economic Immigrants in the ENI", col.names = c("Value", "Count", "Total", "Percent", "Count", "Total", "Percent", "p-value")) %>% 
  kable_styling(latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c("Variable" = 1, "Economic Immigrants" = 3, "Non-economic Immigrants" = 3, "Test of Equal Props." = 1)) %>%
  pack_rows(index = c("Currently working?" = 2, "Gender" = 2, "Level of studies" = 5, "Country classification" = 8, "Top sectors of work" = 9, "Autonomous Community" = 19))
```

I also find that there are significant differences in the gender, educational level, country of origin, and sectoral compositions of the two sample populations. Whereas non-economic immigrants are more likely to be female due to the fact that the majority of women in this group come for family reunification reasons, economic immigrants are more likely to be newer arrivals from developing countries with shorter migratory histories to Spain such as Romania, Ecuador, Colombia, and Bolivia. While out of all those surveyed, many reside in the metropolitan regions of Madrid and Catalunya, there is proportionately more economic immigrants living in Murcia, Navarra, and Aragón than non-economic immigrants. Economic immigrants also tend to be younger on average, have lower rates of tertiary educational attainment, and are more concentrated in the agricultural, construction, and household sectors.^[See Table \ref{tab:tabmeans} in Appendix for distributions of age and years in Spain.]

### Response Variable: Log of hourly wage
Working respondents were asked to report how much money they receive a month for their jobs, including the monthly proportion of any regularly-earned outstanding payments, as well as how many hours a week they work habitually at their jobs. Of the 6,369 economic immigrants that reported their current laboral situation as "working" (74% of all economic immigrants), 5,093 reported their monthly earnings with a real value as well as their weekly worked hours (80% of working economic immigrants, 59% of all economic immigrants).^[Reported monthly earnings were not adjusted for provincial differences in the Consumer Price Index as systematic variation in wages due to location will be accounted for with geographical controls.] Estimated hourly earnings are calculated by dividing total monthly earnings by the estimated total hours worked habitually a month.^[To estimate total hours worked habitually a month, total hours worked habitually a week are multiplied by 4.]

Table \ref{tab:tabwages} shows the average reported hourly wage and sample size for each group from country *c* living in province *p* at the time of surveying. Countries of origin and provinces displayed are those with the most representation among the entirety of the population sample of economic immigrants with reported earnings who arrived in Spain in 1998 or later.^[See Section 2.2.1 for rationale behind this restriction.] While average wage levels differ at both the country-of-origin level and the province level, there are also differences in wage level within groups. For example, while Peruvians earn higher wages than the average economic immigrant, they earn about 12% less than the province-wide average in Zaragoza but 13% more than the province-wide average in Barcelona. Romanians, on the other hand, make about 3% more in Zaragoza but 13% less in Barcelona. Of course, these within-group differences could very well be due to underlying systematic differences in age, education level, socioeconomic background, or sector of work. The empirical analysis shows, however, that these within-group regional differences still remain after controlling for various observable demographic, human capital, and work-related factors.

```{r echo = FALSE, warning = FALSE, message = FALSE}
# Average wages for selected national origin groups in selected provinces
table1 <- this_pop_1 %>% filter(!is.na(hour_earn)) %>% dplyr::select(country, province, hour_earn) %>% filter(country %in% c("Rumanía", "Ecuador", "Colombia", "Marruecos", "Bolivia", "Argentina", "Perú") & province %in% c("Madrid", "Murcia", "Barcelona", "Navarra", "Baleares (Illes)", "Valencia", "Zaragoza", "Rioja (La)")) %>% group_by(country, province) %>% dplyr::summarize(mean_hour_earn = paste0(round(mean(hour_earn), 2), " (", dplyr::n(), ")")) %>% spread(key = country, value = mean_hour_earn)

row_append = c("Total", this_pop_1 %>% filter(!is.na(hour_earn)) %>% filter(country %in% c("Rumanía", "Ecuador", "Colombia", "Marruecos", "Bolivia", "Argentina", "Perú")) %>% group_by(country) %>% dplyr::summarize(mean_hour_earn = paste0(round(mean(hour_earn),2), " (", dplyr::n(), ")")) %>% pull(mean_hour_earn))

column_append = this_pop_1 %>% filter(!is.na(hour_earn)) %>% filter(province %in% c("Madrid", "Murcia", "Barcelona", "Navarra", "Baleares (Illes)", "Valencia", "Zaragoza", "Rioja (La)")) %>% group_by(province) %>% dplyr::summarize(mean_hour_earn = paste0(round(mean(hour_earn),2), " (", dplyr::n(), ")")) %>% pull(mean_hour_earn)

table1 <- table1 %>% rbind(., row_append) %>% cbind(., Total = c(column_append, paste0(round(mean(this_pop_1$hour_earn, na.rm = T),2), " (", nrow(this_pop_1 %>% filter(!is.na(hour_earn))), ")"))) 
```

```{r tabwages, echo = FALSE, warning = FALSE, message = FALSE, fig.height = 7.0, fig.width = 9}
## Output stats
kable(table1, format = "latex", row.names = F, align = 'c', 
      col.names = c("", "Argentina", "Bolivia", "Colombia", "Ecuador", "Morocco", "Peru", "Romania", "Total"),
      caption = "Mean hourly wages for top national origin groups in selected provinces") %>%
  column_spec(1, bold = T, border_right = T) %>%
  column_spec(8, border_right = T) %>%
  row_spec(0, bold = T, hline_after = TRUE) %>%
  row_spec(8, hline_after = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "scale_down")) %>%
  add_footnote(label = c("Sample size of each group shown in parentheses.", "Column-wise means include all provinces and row-wise means include all countries of origin, not just those displayed."),
               notation = "number") 
```

\vspace{-3truemm}
## *Estadística del Padrón continuo* (*Padrón* data) from 1998 - 2007
\vspace{-3truemm}
Aggregate population data by year and Spanish province is taken from the municipal registry database (*Padrón municipal*) in order to map province-level yearly residential population data to immigrants in the ENI by country of origin.^[Of the 112 countries of origin represented in the study population of economic immigrants from the ENI, the *Padrón* data is summarized by exact nationality for 97. Those from Azerbaiyán, Uzbekistán, Mozambique, Tanzania, Chad, Zimbabwe, Trinidad and Tobago, Suriname, Puerto Rico, and Sri Lanka, represented by a total of 16 respondents, are categorized in aggregated groups corresponding to those in the *Padrón* data.] Registry data is available at the province-level (NUTS-3) and at the municipal-level. However, only province-level data will be considered since 1) municipal-level data is only available from 2003 onwards and 2) linking municipal-level population data to individuals in the ENI is not currently possible due to data protection reasons.^[Information on municipality of residence in Spain and history of movements between municipalities is available in the ENI, but fictitious names of the municipalities are assigned for data protection reasons in accordance with the *Secreto Estadístico* of Spain. A fee of 367,84\texteuro must be paid in order to access this data.] ^[Link to the aggregated data files with registry information counted by year, province, and nationality can be found at https://www.ine.es/jaxi/Tabla.htm?path=/t20/e245/p04/provi/l0/&file=0ccaa002.px&L=0] 

Registering in the local municipality system is obligatory for *anyone* who resides in Spain. There are incentives to do so for Spanish and EU nationals as it provides access to public medical care and education. Residency authorizations for foreigners are contingent upon being *empadronado*. Even undocumented immigrants have motives to register themselves as the regularization mechanism of *arraigo*, established in 2004, grants legal status to foreigners who, conditional on various other requirements, can prove that they have been living in Spain for at least two years [@galvez2020size]. Thus, the *Padrón* data is advantageous for the purposes of this study as it provides accurate residential data on the broad range of foreign-born populations living in Spain as of 2007. However, it is these same policy changes and the incentives they create for registering in the system that could introduce great measurement bias into the data, an issue I will turn back to in Section 4.2.1.

### Explanatory Variable: Co-ethnic density
A measure of co-ethnic density assigned to each individual is defined as the proportion of the province's total resident population who are of the same national origin and gender. Co-ethnic matching is done by gender with the understanding that within the segregated labor markets of natives and immigrants, there is further segregation on the lines of gender. In the ENI population, the top-represented sectors of work are highly segregated. For example, the construction sector is 98.7% male and 1.3% female, household activities are 2.6% male, 97.4% female, hospitality 31.0% male, 69.0% female, and manufacturing 71.7% male, 28.3% female. Thus, if co-ethnic density has an effect on the economic lives of immigrants, that effect will be best captured through co-ethnic and co-gender networks. 

The co-ethnic density index can thus be formalized as such:

\begin{equation}
\label{eq:abs_eth_dens}
\text{Co-Ethnic Density} = \frac{N_{c,g,p}}{N_{g,p}} 
\end{equation}

where $N_{c,g,p}$ represents the number of individuals born in foreign country *c* of gender *g* living in Spain in province *p* as of the 1st of January of the year in which the individual of reference reported having begun living in that province. Since *Padrón* data stratified by province and nationality is only available from 1998 onwards, the study population will be restricted to those who arrived to province *p* in 1998 or later.^[Any ENI respondents residing in the autonomous cities of Ceuta and Melilla are excluded from the sample.] $N_{g,p}$ therefore represents the *total* number of residents of same gender *g* living in province *p* in that same year, foreign-born and natives alike.^[There are 24 individuals in the study population who live in provinces where no other co-ethnics are registered as living in in the year the individual started living there. These are principally those who come from small countries with little pre-established migratory history to said provinces. Since the case of no other co-ethnic residents living in such a geographically large area is improbable and since I know that at least the one ENI respondent was reportedly living in the province in said year, I impute the missing numerator $N_{c,g,p}$ with a value of one.]

```{r echo = FALSE, warning = FALSE, message = FALSE}
# Average wages for selected national origin groups in selected provinces
table2 <- this_pop_1 %>% filter(!is.na(eth_dens_prov_1) & !is.na(hour_earn)) %>% dplyr::select(country, province, eth_dens_prov_1) %>% filter(country %in% c("Rumanía", "Ecuador", "Colombia", "Marruecos", "Bolivia", "Argentina", "Perú") & province %in% c("Madrid", "Murcia", "Barcelona", "Navarra", "Baleares (Illes)", "Valencia", "Zaragoza", "Rioja (La)")) %>% group_by(country, province) %>% dplyr::summarize(mean_eth_dens_prov_1 = mean(eth_dens_prov_1*10), sd_eth_dens_prov_1 = sd(eth_dens_prov_1*10)) %>% mutate(
  result = paste0(round(mean_eth_dens_prov_1, 2), " (", round(sd_eth_dens_prov_1, 1), ")")) %>%
  dplyr::select(-c(mean_eth_dens_prov_1, sd_eth_dens_prov_1)) %>% spread(key = country, value = result)

row_append = c("Total", this_pop_1 %>% filter(!is.na(eth_dens_prov_1)) %>% filter(country %in% c("Rumanía", "Ecuador", "Colombia", "Marruecos", "Bolivia", "Argentina", "Perú")) %>% group_by(country) %>% dplyr::summarize(mean_eth_dens_prov_1 = paste0(round(mean(eth_dens_prov_1*10),2), " (", round(sd(eth_dens_prov_1*10), 1), ")")) %>% pull(mean_eth_dens_prov_1))

column_append = this_pop_1 %>% filter(!is.na(eth_dens_prov_1)) %>% filter(province %in% c("Madrid", "Murcia", "Barcelona", "Navarra", "Baleares (Illes)", "Valencia", "Zaragoza", "Rioja (La)")) %>% group_by(province) %>% dplyr::summarize(mean_eth_dens_prov_1 = paste0(round(mean(eth_dens_prov_1*10),2), " (", round(sd(eth_dens_prov_1*10), 1), ")")) %>% pull(mean_eth_dens_prov_1)

table2 <- table2 %>% rbind(., row_append) %>% cbind(., Total = c(column_append, paste0(round(mean(this_pop_1$eth_dens_prov_1*10, na.rm = T),2), " (", round(sd(this_pop_1$eth_dens_prov_1*10, na.rm = T),1), ")"))) 
```

```{r tabethdens, echo = FALSE, warning = FALSE, message = FALSE, fig.height = 7.0, fig.width = 9}
## Output stats
kable(table2, format = "latex", row.names = F, align = 'c', 
      col.names = c("", "Argentina", "Bolivia", "Colombia", "Ecuador", "Morocco", "Peru", "Romania", "Total"),
      caption = "Mean co-ethnic density per 1,000 residents for top 7 most represented national origin groups in selected provinces") %>%
  column_spec(1, bold = T, border_right = T) %>%
  column_spec(8, border_right = T) %>%
  row_spec(0, bold = T, hline_after = TRUE) %>%
  row_spec(8, hline_after = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "scale_down")) %>%
  add_footnote(label = c("Standard deviations shown in parentheses.", "Column-wise means include all provinces and row-wise means include all countries of origin, not just those displayed."),
               notation = "number")

```

The column-wise variability in co-ethnic density we see in Table \ref{tab:tabethdens} shows that immigrants originating from the same country choose to settle in different provinces at different times, but that these geographical settlement preferences are not necessarily similar across groups (rows), providing evidence that co-ethnic networks play a role in the settlement patterns of immigrants in Spain. The empirical analysis that follows investigates how these networks may continue to play a role in the economic lives of immigrants once settled. 

# METHODOLOGY {#methodology}
To test whether co-ethnic density, $COED$, has an effect on the log of hourly wage, $logWAGE$, of immigrant $i$, I will utilize the following theoretical model: \begin{equation} \label{eq:wage_equation}
log(WAGE_{i}) = \beta_0 + \beta_1COED_{i} + \boldsymbol{\rho X_i} + \zeta SECT_i + \gamma COUNTRY_i + \psi REG_i + \epsilon_i
\end{equation} where $\bf{X_i}$ denotes a vector of individual-level covariates capturing level of education attained and the number of years since migration as well as demographic variables for age, the square of age and gender of the individual.^[Level of completed education is reported on the ENI as one of the following categories: "No studies", "Some studies", "Primary studies (incomplete)", "Primary studies (complete)", "Secondary studies (1st cycle)", "Secondary studies (2nd cycle)", "Tertiary studies (1st cycle)", and "Tertiary studies (2nd cycle)". While the use of *years* of completed education is common when studying the return to education in the United States context (see @card1999causal for the relevant discussion), I opt to maintain categorical factors representing completed levels of education in the following hierarchical order: "No secondary studies", "Secondary studies (1st cycle)", "Secondary studies (2nd cycle)", "Tertiary studies (1st cycle)", and "Tertiary studies (2nd cycle)". Public education in Spain is free and obligatory up until the age of 16, the year in which an individual can stop studying or continue his secondary studies in the 2nd cycle (*Bachillerato* or *formación profesional de Grado Medio*). Therefore, it is unlikely that differences in schooling level for those who do not have any secondary education would have a substantially different impact on wages in the Spanish context.] ^[According to the seminal work of @chiswick1978, when measuring the earnings of immigrants individual-level human capital can be proxied for by time spent in the destination country as one goes assimilating to the new culture, job market, norms, language, and general way of life in which he finds himself.] ^[Information on birth year and arrival year was missing for 28 individuals. Values for age and number of years since migration were imputed for these individuals via conditional mean imputation, regressing them individually on all other independent variables in the outcome model for the set of observations with complete information and using the model to predict the missing values. See Appendix for specifications of the imputation model.] $SECT_i$ denotes a control for sector of work to unconfound any co-ethnic residential clustering that may be a result of sectoral regionalization such as in the agricultural or manufacturing industries, the former localized in the less dense southern provinces in Andalucía and Murcia while the latter in the more concentrated regions of the north.^[A series of bilateral agreements reached in the early 2000s expanded work visa quotas in the construction sector coming from some of the largest immigrant exporting countries at the time like Ecuador, Colombia, Dominican Republic, and Morocco highlighting how regional sectors can influence observed co-ethnic clustering [@imopolicies].] Finally, $COUNTRY_i$ and $REG_i$ denote vectors of dummy variables indicating country of origin and region of residence, respectively. Region-level fixed effects allows for differing wage levels between geographical regions that affect all groups within the region similarly due to general differences in cost-of-living or labor market conditions, a prominent feature of the Spanish system (see @maza2009provincial and @liu2018regional for the relevant discussion). Country-of-origin fixed effects, on the other hand, controls for factors that arguably affect all co-ethnic individuals similarly when it comes to economic integration and assimilation in a foreign country, such as language, culture, or institutions. Regions are stratified within provinces by population, and while this does not account for the geographical size of the municipality, the distinction between urban and rural regions of each province generally correspond to the population stratifications utilized, this relationship visualized in Figure \ref{fig:figmun} of the Appendix where one can see high population count municipalities clustered around the major metropolitan areas. Finally, to avoid singularities, I restrict the population to the top 43 most represented countries of origin and top 111 most represented regions.^[For each grouping, $n > 10$.] 

Since estimating the effect of co-ethnic density on wages is only possible to do for those working and hence generating income, I use Heckman's sample selection model to test whether there are any systematic differences between the populations of economic immigrants who reported their wages versus those who did not, understanding that these differences could bias my estimates [@heckman1979sample]. For example, immigrant women whose cultures restrict them from working outside the home, often those that also restrict access to education and other basic rights for women, may be more likely to reside in highly dense co-ethnic areas. These women, if they could work, would on average make lower wages, so by not observing them the estimates for the effect of co-ethnic density on wages may be biased upwards. Comparing the two populations in Table \ref{tab:sampselection} of the Appendix, we see that women are less likely to have been observed as working than men are, particularly for those women who are married with children living in the same household. We also see that those from Andean countries and Eastern Europe, the Balkans, and other European nationalities are more likely to be working whereas those from developed countries and northern Africa & the Middle East are less likely to be working. Given these results, I specify the selection equation such that the probability of having wage data depends on whether one is female or not, what country classification his or her country of origin belongs to (classifications shown in Table \ref{tab:countrygroup}), and whether or not he or she is a married woman with children in the house. Regressing a Probit model on the selection equation, I calculate the inverse Mills ratio, $\hat{\lambda}$, and estimate the wage model with $\hat{\lambda}$ as an added regressor. I find correlation between the error terms of the selection and outcome equations in all specifications of the wage model such that I reject $H_0: \hat{\lambda} = 0$ and claim that there is significant reason to believe that there is a problem of sample selection when estimating my wage model.^[Significance level determined at $\alpha \le 0.05$] Therefore, all results shown in Section \ref{results} are based on the Heckman selection model. 

```{r echo = FALSE, warning = FALSE, message = FALSE}
this_pop_economic_wage <- this_pop_1 %>% filter(wage_data == 1)
this_pop_economic_no_wage <- this_pop_1 %>% filter(wage_data == 0)

prop_test_sampselection <- prop_test_function(this_pop_1 %>% filter(wage_data == 1), this_pop_1 %>% filter(wage_data == 0), var_list = c("sex", "education_eng_class", "country_class", "woman_married_kids"))
prop_test_sampselection$var[which(prop_test_sampselection$var %in% c(0,1))] <- mapvalues(prop_test_sampselection$var[which(prop_test_sampselection$var %in% c(0,1))], c(0, 1), c("No", "Yes"))
prop_test_sampselection <- prop_test_sampselection[, -c(1)]

comp_table_sampselection <- rbind(
  cbind(var = "Years in Spain", 
        mean_wage = round(mean(this_pop_economic_wage$years_final, na.rm = T), 2),
        sd_wage = round(sd(this_pop_economic_wage$years_final, na.rm = T), 2),
        n_wage = sum(!is.na(this_pop_economic_wage$years_final)),
        mean_no_wage = round(mean(this_pop_economic_no_wage$years_final, na.rm = T), 2),
        sd_no_wage = round(sd(this_pop_economic_no_wage$years_final, na.rm = T), 2),
        n_no_wage = sum(!is.na(this_pop_economic_no_wage$years_final)),
        p = signif(t.test(this_pop_economic_wage$years_final, this_pop_economic_no_wage$years_final, alternative = "two.sided", var.equal = F)[3]$p.value, 2)),
  cbind(var = "Age", 
        mean_wage = round(mean(this_pop_economic_wage$age_final, na.rm = T), 2),
        sd_wage = round(sd(this_pop_economic_wage$age_final, na.rm = T), 2),
        n_wage = sum(!is.na(this_pop_economic_wage$age_final)),
        mean_no_wage = round(mean(this_pop_economic_no_wage$age_final, na.rm = T), 2),
        sd_no_wage = round(sd(this_pop_economic_no_wage$age_final, na.rm = T), 2),
        n_no_wage = sum(!is.na(this_pop_economic_no_wage$age_final)),
        p = signif(t.test(this_pop_economic_wage$age_final, this_pop_economic_no_wage$age_final, alternative = "two.sided", var.equal = F)[3]$p.value, 2))) %>% as.data.frame()

```

When analyzing the effect of co-ethnic density on the wages of immigrant *i*, the coefficient of interest, $\beta_{1}$, represents the difference in the log of hourly wage, or the percentage change in hourly wage, if that same person were to have settled in a province with one percentage point more of the population made up of co-ethnics. Due to the lack of theoretical understanding of the mechanisms at work and the contradicting empirical evidence about the directionality of the effect as discussed previously, I know turn to the data to see if there is any independent effect of co-ethnic density on immigrant wages. 

# RESULTS {#results}

## Estimating Effect of Co-Ethnic Density via OLS
Table 4 presents the results from the Heckman sample selection wage model using various specifications of Equation \ref{eq:wage_equation} to estimate the effect of co-ethnic density, $\beta_1$ on hourly earnings of economic immigrants. Models 1 through 8 show the results of the adjusted OLS regressions, progressively controlling for more covariates with each iteration of the model. I expect substantial multicollinearity to exist between my explanatory variable of interest and the other covariates in the model given the fact that residential location, in the case of my sample, is not determined randomly. Indeed, the initial negative effect of co-ethnic concentration seen in Model 1, representing about a 1.1% decrease in monthly earnings for every one percentage point increase in co-ethnic density, increases in magnitude by almost three when controlling for gender of the individual indicating that males, who on average make significantly more than their female counterparts, also tend to settle in more relatively dense co-ethnic provinces. Those who settle in more relatively dense co-ethnic provinces also tend to be newer immigrants with lower education level who make less on average than their more established and educated counterparts, as seen by the change in the estimated effect of co-ethnic density when controlling for years in Spain in Model 3 and educational achievement in Model 4. However, the independent effect of co-ethnic density on wages remains negative and significant even after the addition of controls for age, the square of age, and sector of work. The additional fixed effect for country of origin (Model 7) shifts the co-ethnic density effect estimate slightly upwards, indicating that those from certain countries of origin who have lower earning levels as a whole tend to settle in more concentrated provinces. Finally, the addition of the fixed effect for region of residence in the fully specified model (Model 8) does not change substantially the results. While insignificant at $\alpha = 0.1$, we do reject the null hypothesis that $COED = 0$ at $\alpha = 0.14$

To interpret concretely what the results from Model 8 suggest, we may understand the effect of co-ethnic density on wages as such: the wage of individual *i* in a parallel universe would be about 1% greater if the only thing that had changed about her situation was that she had decided to settle in a province with one percentage point less of the residents hailing from her home country. This effect is independent of region, which is perhaps most surprising given that I expect certain metropolitan areas to attract more immigrants to them due to the fact that they have higher wage levels and better employment opportunities. If this is indeed the case, it suggests that the true effect of co-ethnic density may be even more negative. 

```{r echo = FALSE, warning = FALSE, message = FALSE}
# Restrict to primary study population 
######### 1. Effect of concentration of coethnics in same province on monthly earnings
model_data_1i_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1,
                               method = "2step",
                               data = this_pop_1)

model_data_1ii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1 + sex,
                               method = "2step",
                               data = this_pop_1)

model_data_1iii_adj <- selection(wage_data ~ sex + country + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1 + sex + years_final,
                               method = "2step",
                               data = this_pop_1)

model_data_1iv_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1 + sex + years_final + education_eng_class,
                               method = "2step",
                               data = this_pop_1)

model_data_1v_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1 + sex + years_final + education_eng_class + age_final + age_sq_final,
                               method = "2step",
                               data = this_pop_1)

model_data_1vi_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1 + sex + years_final + education_eng_class + age_final + age_sq_final + sector2,
                               method = "2step",
                               data = this_pop_1)

model_data_1vii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1  + sex + years_final + education_eng_class + age_final + age_sq_final + sector2 + country,
                               method = "2step",
                               data = this_pop_1)

model_data_1viii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ eth_dens_prov_1 + sex + years_final + education_eng_class + age_final + age_sq_final + sector2 + country + province_strata,
                               method = "2step",
                               data = this_pop_1)
```

```{r tabresults1, echo = FALSE, warning = FALSE, message = FALSE, results='asis', resize.height=4,resize.width=11}
#### WAGE TABLE (ABSOLUTE ETHNIC DENSITY) ####
  stargazer(model_data_1i_adj, model_data_1ii_adj, model_data_1iii_adj, model_data_1iv_adj, model_data_1v_adj,
                                      model_data_1vi_adj, model_data_1vii_adj, model_data_1viii_adj, 
          type="latex",
          title = "Effect of co-ethnic density on the log of hourly wage",
          colnames = F,
          dep.var.labels = "Log of hourly wage",
          dep.var.labels.include = T,
          covariate.labels = c("COED",  "Female", "Years in Spain" , "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)", "Age", "Age squared"),
          keep = c("eth_dens_prov_1", "sex", "years_final", "education_eng_class", "age_final", "age_sq_final"),
          omit = c("sector2", "country", "province_strata"),
          add.lines=list(c('Sector controls', rep("No", 5), rep("Yes", 3)),
                         c('Origin country controls', rep("No", 6), rep("Yes", 2)),
                         c('Region controls', rep("No", 7), rep("Yes", 1))),
          column.sep.width = "1pt",
          #column.labels   = c("Absolute co-ethnic density", "Relative co-ethnic density"),
          #column.separate = c(4, 4),
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = F,
          font.size = "scriptsize",
          notes = c("All results shown are based on the Heckman selection model."),
          notes.align = "l",
          keep.stat = c("n", "adj.rsq",  "f", "ser", "chi2"),
          table.placement = "H")

```

\vspace{-3truemm}
## Estimating Effect of Co-Ethnic Density (Absolute and Relative) via 2SLS {#ivreg}
The results presented above suggest that the overall effect of residing in a highly dense co-ethnic area is bad for immigrant wages. However, it does not necessarily shed light on the mechanisms behind what may very well be the net impact of various positive and negative forces. The theory that ethnic enclaves, especially linguistic ones, can reduce rates of interaction with the native population as well as the need to improve foreign language ability or assimilate other foreign-acquired human capital has found empirical support: @barry2005enclaves show that ethnic clustering in the United States is associated with lower earnings amongst male immigrants from non-English speaking countries along with lower English attainment, while @kanas2012social provide evidence that residential co-ethnic concentration in Germany has a weak negative effect on the economic returns of education achieved in the host country. 

However, research on immigrant enclaves has spurred heated debate on how to operationalize such a complex phenomenon [@portes2007revisiting]. Therefore, I shall take extra care to emphasize that I am *not* measuring the effect of ethnic enclaves on economic outcomes of immigrants, and indeed the explanatory variable of interest I have employed thus far is very far from being able to identify these highly localized and contextualized sociocultural and economic relationships. While my measure of co-ethnic density as defined above may be capturing the effect of networks that arise from living in relatively dense co-ethnic areas, it could also represent, and perhaps more intuitively represents, the magnitude of immigrant inflows given that the immigrant share of the total population during the study period was growing at a very fast rate at the time (Figure 1). If this is the case, then it would make sense that wages for immigrants are depressed in areas where there are greater stocks of immigrant labor since greater job competition arises under the assumption that arriving immigrant labor is a close substitute for that of already arrived ones [@borjas1997much]. Since upward mobility is difficult for the otherwise qualified immigrant who faces legal, institutional, and language barriers to accessing the broader job market, more and more get filtered into the more easily accessible low-paying secondary sectors where less specialization and human capital assimilation is required, the same sectors that were in search of cheap labor during this sustained period of prosperity and growth in the Spanish economy. In fact, this is the theory employed in other empirical studies that have shown high rates of over-education in certain immigrant populations [@joona2014overeducation] [@hou2021trends].

With this argument in mind, I opt to include a measure of *relative co-ethnic density* into Equation \ref{eq:wage_equation}, which adjusts the co-ethnic density index from Equation \ref{eq:abs_eth_dens}, now referred to as the *absolute co-ethnic density*, for each individual from country *c* of gender *g* living in province *p* by the absolute co-ethnic density index of all individuals from country *c* of gender *g* living in Spain such that:

\begin{equation}
\label{eq:rel_eth_dens}
\text{Relative Co-Ethnic Density} = \frac{\frac{N_{c,g,p}}{N_{g,p}}}{\frac{N_{c,g}}{N_g}}
\end{equation}

where $N_{c,g}$ represents the number of individuals born in foreign country *c* of gender *g* living in Spain in the year of the individual of reference's arrival in province *p* and $N_{g}$ represents the number of all individuals of gender *g* living in Spain in that same year. In this way, even if the overall counts for some of the cohorts are small, the relative co-ethnic density index will be more able to capture differences in network *strength*, whereas the absolute measure of co-ethnic density is better at capturing network *magnitude*.  

To make this idea more concrete, in 2004, immigrants from Cabo Verde made up only about 0.006% of the total Spanish resident population. That means that for every 100,000 residents in Spain there were 6 from Cabo Verde. However, there were nearly 100 Cabo Verdeans for every 100,000 residents registered in León at the time. That means that Cabo Verdeans were over sixteen times more concentrated in León than if they were to perfectly disperse themselves throughout any of the other provinces. The size of this effect for what is a relatively low-represented immigrant group in absolute terms suggests a specific quality about the migratory networks that were leveraged in the settlement patterns of Cabo Verdeans in Spain, and I hypothesize that these networks may continue having an effect on the economic lives of immigrants once settled. 

Figure \ref{fig:figdenscomp} illustrates the differences between the two indices for selected countries of origin. While the absolute measure of co-ethnic density index is able to capture patterns of residential clustering for those countries of origin with longer or more intense migratory histories like Morocco or Romania, it is only with the relative measure of co-ethnic density that these patterns emerge for countries with lower magnitude or more recent migratory flows like from Cabo Verde or Venezuela.

```{r echo = FALSE, warning = FALSE, message = FALSE}
## Which countries are most represented?
countries <- table(this_pop_1$country) %>% as.data.frame(responseName = "Freq") %>% 
  dplyr::rename(country = Var1) %>% arrange(desc(Freq))
these_top_countries <- countries %>% filter(Freq > 300)

####################    Padrón directory specification   #######################
padron_prov_data_name <- "poblacion_provincia_serie_temporal.xlsx"

####################   Prepare and merge geo data #########################
## Population density attributes by province and country of origin
dat_padron_prov <- read.xlsx(paste0(data_dir, "/", padron_prov_data_name), sheet = 3, startRow = 8, cols = c(3:28)) %>% 
  dplyr::select(c(Destination_mapSpain, Origin, `1998`:`2021`)) %>% melt(., id.vars = c("Destination_mapSpain", "Origin")) %>%
  dplyr::rename(eth_dens_prov = value, year = variable) 

dat_padron_prov <- left_join(x = dat_padron_prov %>% filter(Destination_mapSpain != "Total"),
                  y = dat_padron_prov%>% filter(Destination_mapSpain == "Total") %>% dplyr::select(-c("Destination_mapSpain")), 
                  by = c("year", "Origin"),
                  suffix = c("", "_Total"),
                  keep = F) %>%
  mutate(eth_dens_prov_relative = eth_dens_prov/eth_dens_prov_Total,
         log_eth_dens_prov_relative = log(eth_dens_prov_relative))

## Get province shapes
provinces <- esp_get_prov_siane()

dat_geo_prov_comp <- sp::merge(provinces,
                    dat_padron_prov %>% filter(Origin %in% c("Marruecos", "Rumanía", "Cabo Verde", "Venezuela") & year == "2004") %>% dplyr::select(Destination_mapSpain, Origin, eth_dens_prov, eth_dens_prov_relative) %>% melt(),
                    by.x = "ine.prov.name",
                    by.y = "Destination_mapSpain",
                    all.x = TRUE
) 

################### Plot with tmap #####################
## Population density in 2006 for each country of top origin
ethnic_density_top_2004 <- tm_shape(dat_geo_prov_comp) +
  tm_polygons("value", palette = "Reds", style="cont", 
              title = "Co-Ethnic Density",
              id = "ine.prov.name",
              breaks = c(0, 1, 2, 5, 10, 15, 20),
              legend.is.portrait = FALSE) +
  tm_facets(by = c("variable", "Origin"), free.coords = FALSE, nrow = 1, ncol = 6) +
  tm_layout(legend.outside = TRUE,
            legend.outside.position = "bottom",
            asp=0,
            panel.label.size = 2,
            panel.labels = list(c("Absolute", "Relative"),
                                c("Cabo Verde", "Morocco", "Romania", "Venezuela")
                                )) +
  tm_credits(text = c(rep("", 7), "Data source: INE's padrón continuo"), size = 1.5)
```

```{r figdenscomp, echo = FALSE, warning = FALSE, message = FALSE, results='hide', fig.keep='all', fig.align = "center", fig.height = 8, fig.width = 10, strip.white=T, fig.cap = "Absolute and Relative Co-Ethnic Density Measures for Select Countries of Origin (2004)"}
ethnic_density_top_2004
```

When estimating the relative co-ethnic density index on wages, I take the log of the index so that any negative value represents province co-ethnic densities that are less than the national density, which we may interpret as representing weaker networks, and positive values represent those with relatively more co-ethnic concentrations than the national average, representing stronger networks. I also log the measure of absolute co-ethnic density to achieve more variance in the distribution at lower domain values and control for this such that my new wage equation looks as such:
\begin{equation} \label{eq:new_wage_equation} \begin{aligned}
log(WAGE_{i}) = \beta_0 + \beta_1log(absCOED_{i}) + \beta_2log(relCOED_{i}) +  \boldsymbol{\rho X_i} + \zeta SECT_i + \\ \gamma COUNTRY_i + \psi REG_i + \epsilon_i
\end{aligned}
\end{equation}

We can see in Table 5 that the estimated effect for the log of relative co-ethnic density throughout all specifications of the model is positive and quite significant. Remembering that in this analysis absolute co-ethnic density is logged, we still see the same general pattern of a negative and significant effect as we saw in Table 4. The patterns of how the effect changes with each added covariate are also generally the same. It is interesting to note, however, that when controlling for country of origin in Model 7, the effect of absolute density becomes more negative unlike in Model 7 of Table 4 where it becomes more positive. This positive effect instead appears to "move" to the relative density effect. The shift from Model 6 to Model 7 in Table 4 indicates that those from certain countries of origin that have lower baseline wages are also those that have high measures of absolute density, or, as discussed before, greater overall magnitude of co-ethnics living there. However, when we control for differences in relative density, or the likelihood of settling in one area over another *within* ethnic groups, we see the negative effect of the former become more negative while the positive effect of the latter become more positive. In this way, by controlling for both absolute and relative ethnic density, the "stock" versus "strength" effect of migrant networks disentangle themselves. These results suggest that for a single immigrant it is more economically advantageous to live in a province with low absolute but high relative co-ethnic density, or more plainly put, within a strong but small co-ethnic network. 

```{r echo = FALSE, warning = FALSE, message = FALSE}
## ADDITION OF RELATIVE ETHNIC DENSITY
model_data_relativei_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative,
                               method = "2step",
                               data = this_pop_1)

model_data_relativeii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex,
                               method = "2step",
                               data = this_pop_1)

model_data_relativeiii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex + years_final,
                               method = "2step",
                               data = this_pop_1)

model_data_relativeiv_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex + years_final + education_eng_class,
                               method = "2step",
                               data = this_pop_1)

model_data_relativev_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex + years_final + education_eng_class + age_final + age_sq_final,
                               method = "2step",
                               data = this_pop_1)
model_data_relativevi_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex + years_final + education_eng_class + age_final + age_sq_final + sector2,
                               method = "2step",
                               data = this_pop_1)

model_data_relativevii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex + years_final + education_eng_class + age_final + age_sq_final + sector2 + country,
                               method = "2step",
                               data = this_pop_1)

model_data_relativeviii_adj <- selection(wage_data ~ sex + country_class + woman_married_kids,
                               log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + sex + years_final + education_eng_class + age_final + age_sq_final + sector2 + country + province_strata,
                               method = "2step",
                               data = this_pop_1)

```

```{r tabresults2, echo = FALSE, warning = FALSE, message = FALSE, results='asis', resize.height=4,resize.width=11}
#### WAGE TABLE (ABSOLUTE and RELATIVE ETHNIC DENSITY) ####
  stargazer(model_data_relativei_adj, model_data_relativeii_adj, model_data_relativeiii_adj, model_data_relativeiv_adj, model_data_relativev_adj, model_data_relativevi_adj, model_data_relativevii_adj, model_data_relativeviii_adj,
          type="latex",
          title = "Effect of log of co-ethnic density (absolute and relative) on log of hourly wage",
          colnames = F,
          dep.var.labels = "Log of hourly wage",
          dep.var.labels.include = T,
          covariate.labels = c("logCOED (absolute)", "logCOED (relative)", "Female", "Years in Spain" , "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)", "Age", "Age squared"),
          keep = c("log_eth_dens_prov_1", "log_eth_dens_prov_1_relative", "sex", "years_final", "education_eng_class", "age_final", "age_sq_final"),
          omit = c("sector2", "country", "province_strata"),
          add.lines=list(c('Sector controls', rep("No", 5), rep("Yes", 3)),
                         c('Origin country controls', rep("No", 6), rep("Yes", 2)),
                         c('Region controls', rep("No", 7), rep("Yes", 1))),
          column.sep.width = "1pt",
          #column.labels   = c("Absolute co-ethnic density", "Relative co-ethnic density"),
          #column.separate = c(4, 4),
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = F,
          font.size = "scriptsize",
          notes = c("All results shown are based on the Heckman selection model."),
          notes.align = "l",
          keep.stat = c("n", "adj.rsq", "f", "ser", "chi2"),
          table.placement = "H")

```

### Endogeneity
Given its spatial nature, co-ethnic density as defined above would be a more accurate measure of network effects if it could be measured at a more disaggregate level. However, as mentioned previously, information on the municipality of residence is anonymized on the ENI data making it impossible to link respondents to the municipality-level *Padrón* data. While maintaining my measure of residential co-ethnic density at the provincial level complicates the interpretation of the mechanism behind the effect that co-ethnic density may have on economic outcomes, analyzing the variation in co-ethnic population size across a broader geographical region helps alleviate obvious endogeneity concerns that may arise from what is known as the residential sorting problem [@duncan1957measurement]. Settlement patterns of migrants are hardly ever determined randomly, and it is likely that one engages with certain migratory networks or moves to a specific location for different unobserved reasons. For example, those who lack legal working rights or human capital in the form of education or language skills may opt for the protection of an ethnic enclave where they expect these factors to matter less for their economic well-being. Pre-established migration links and family reunification also lead to the filtering of migrants from specific backgrounds to certain destinations. Thus, the clustering of similar types of people in the same spaces may bias estimates of the effect of co-ethnic density on earnings in function of the unobserved "quality" of the immigrant group being referenced [@schuller2022ethnic] [@kanas2012social]. I assume, however, that the residential sorting effect is essentially diluted as the level of covariance between the distribution of unobserved cohort "quality" and the distribution of co-ethnic density is reduced when measuring co-ethnic density at an aggregated geographic level as opposed to at a highly localized level.^[Measuring at the province-level also prevents the calculation of outliers that may happen when measuring co-ethnic density in sparsely populated municipalities.] In fact, this is one argument employed in @cutler1997ghettos famous paper, *Are Ghettos Good or Bad?*, for why the authors decide to measure the impact of racial segregation across American cities rather than within them.

Indeed general differences in motives for migration are already taken into account by restricting the study population to the subset of economic immigrants in the ENI, but this does not quell all of the endogeneity concerns that arise from initial settlement decisions and uneven cohort effects. The migratory network itself becomes an actor in the integration and settlement of new cohorts of immigrants, sorting and distributing groups who choose to engage with the network for various different underlying reasons into different locations. To illustrate this and how it might be affecting the results seen in Table 4 independent of the labor stock argument already discussed at the beginning of this section, I return to the concept of chain migration. To oversimplify what merits an entire discussion that is out of the scope of this paper, for sake of argument I distinguish between the types of individuals that participate in these migratory chains as "leaders" and "followers", the former representing the early pioneers that first set sail for Spain, or at least the first that found success in Spain, who, via social and economic networks, opened the doors for many family members, friends, acquaintances, community members, and fellow countrymen to follow. The selectivity that arises from this distinction is apparent and well documented in the literature.^[See @faist2000, @lindstrom2010pioneers, and @wessendorf2018pathways for some examples.] Pioneers tend to be younger, single men more able to get up and leave and find employment than, say, their female counterparts who have children and face greater restrictions in the labor market. Pioneers may also have more resources to make the trip over and find housing as migratory networks are not yet in place to assist with that, or may generally be more adaptable as the option of finding emotional comfort in an ethnic enclave does not exist yet. If this is the case, it suggests that early or small migration cohorts who may represent this positive selection bias have more tendency to concentrate than those who arrive as part of the chain that are more numerous and diffuse. 

In addition, motivational differences between year-of-arrival cohorts may also have resulted in a type of measurement bias in the data. Up until the late 1980s, Spain had no sort of law regulating immigration as it had never been a major destination for international migrants previously (Hooper, 2019). Expired contracts, bureaucracy, and growing unmatched demand for visa allotments in the construction, agricultural, and hospitality sectors drove up the number of irregular immigrants and in 2004 the regularization mechanism of *arraigo social* was established, granting legal status to foreigners who have been living in Spain for at least two years as long as they can prove that they have worked six months illegally in the country, or that they have Spanish family ties or have socially integrated in a sufficient manner [@galvez2020size].^[Achieving *arraigo* through the latter route requires proof of a work-contract offer that should last at least a year.] The incentives that this provides for registering in the local municipal system regardless of legal status (access to free medical care and public education is also conditional on being registered in the local municipal system), while improving authorities’ and researchers' capacity to paint a more accurate picture of the immigrant situation in Spain, could also represent a change in underlying characteristics between those observed in the *Padrón* data prior to 2004 and those observed after. In this sense, a Cabo Verdean that settled in a province prior to 2004 with relatively higher density of Cabo Verdeans may indicate that the high relative co-ethnic density is actually just measuring areas where there are more established Cabo Verdeans who, say, may have legal residency or work permits and thus were not deterred from registering in the municipal system. Therefore, there may be other provinces with high relative densities of Cabo Verdeans, but these are not observed in the data prior to 2004 if, say, these groups were made up of disproportionately more illegal immigrants and therefore more disadvantaged.^[See @farelo2008cape for an intriguing and complete account of Cape Verdean migration in Spain.] Since return migration is of course an event that takes place with greater probability over time for certain immigrants over others, this too presents a possible selection bias in the data. However, very little can be done to address this issue.

Under this hypothesis, it could be the case that the measure of the effect of relative co-ethnic density on immigrant wages is biased upward if there is a positive relationship between the log of relative co-ethnic density and whether one settled in the province before 2004. However, when testing empirically this relationship, there is insufficient evidence to say that there is any significant effect of having arrived pre-2004 on the log of relative co-ethnic density (t = 1.37, p = 0.17). There is, however, a positive correlation between the logs of absolute and relative co-ethnic density (r = 0.54, 95% CI = [0.52, 0.55]), and if I want to make any claims about the combined effect, $\beta_1 + \beta_2$, on the wages of someone who, say, has a small network as represented with the absolute measure, but a strong network as represented with the relative measure, then measurement bias could be a cause of concern since we know already that absolute co-ethnic density is positively correlated with year of arrival. 

### IV Regression
To summarize all of the above, endogeneity in my study as specified thus far is inevitable. Some studies have exploited the pseudo-random variation in co-ethnic concentration due to dispersion policies that have been implemented for migrant refugees [@damm2009ethnic] [@tumen2015use]. However, in the case of unforced migration such quasi-experimental design does not exist and researchers must find an exogenous source of variation to use if they wish to identify the causal effect of co-ethnic density on immigrant wages. In the face of the non-randomness of networks and settlement patterns, I will instrument my co-ethnic density measures with their corresponding measures calculated from national census data from 1991. Just like the *Padrón* data, the census data provides total population counts by province as well as within-province counts stratified by country of origin from which densities are derived. 

Two conditions must be met in order to achieve consistent estimates via instrumental variable regression. The first is that the instrument must be correlated with the endogenous explanatory variable at hand. With the rapid modernization of the economy from the end of the Franconian regime into the 1980s came the expansion of urban industrial areas throughout the country and a growing concentration of people in Madrid, Barcelona, and other major metropolitan regions situated around the perimeters of the country [@gormsen1986recent]. Naturally, many immigrants at the time also gravitated to these regions and continued to do so during the economic boom from 1998 to 2007. A comprehensive multivariate cluster analysis by @parreno2021mapping shows that most immigrants, regardless of country-of-origin, gravitate towards the bigger urban municipalities in Spain. Therefore, I expect at least some correlation between the co-ethnic density measures from the *Padrón* data and the 1991 census data, driven mainly by the fact that immigrants in general will tend to settle in cities where jobs are, regardless of any pre-existing migratory chains that may have led them there. First stage regression modelling in Models 2 and 3 of Table 6 shows that there is indeed strong correlations between the density measures in 1991 and their respective measures in the ENI population a decade later. 

The second condition that must be fulfilled is that the instrument, conditional on the other exogenous variables in the model, is not correlated with the error term. While there is a plethora of endogeneity issues that may be putting the instrument at risk of the same concerns, the most immediate issue is that of selection bias. As discussed in the previous section, it is likely that immigrants self-select into networks of varying quality, and these networks then distribute them to certain regions over others where they experience certain economic outcomes not based on where they settle, but rather based on differences in motives, preferences, or ability that selected them into those networks in the first place. In 1991, however, there were very few foreigners living in Spain, let alone any chain migratory networks that had been established at that point. For much of the 20th century, Spain's migration outflows outpaced its inflows. Prior to 1974, many Spaniards were leaving the country to work abroad under the guest-worker program that was implemented to fill the labor shortages during the post-war boom in Western Europe [@castles1986guest]. Emigration rates dropped during the next decade while foreign inflows kept a slow but steady pace, maintaining an almost net-zero migration rate between the years 1976 and 1988. By 1990, still only 2% of the Spanish population was made up of foreign nationals, many of whom were spending a temporary amount of time in Spain as they transitioned to more popular immigrant destinations such as France or Germany [@perez2003spain]. Immigrant inflow to Spain at this point was clearly not anywhere close to the scale that it would see in the 2000s. 

However, the intersection of a global economic crisis, closing of borders of long-time immigrant importing countries, strategic geo-economic positioning, and recent admittance into the European Union turned Spain from a "waiting room" into a new migrant destination at the start of the 1990s. With the signing of the Treaty of Accession came also the enactment of the country's first national immigration law, the *Ley de extranjería* of 1986. However, the policy lacked teeth as it viewed immigration as a temporary issue, something that needed to be done to deal with those foreigners already living in Spain in the face of the newly minted EU membership card. The 1996 amendment of the national immigration law signaled the recognition by the Spanish government of immigration as a structural phenomenon. Liberalized regularization procedures, extension of immigrant rights, and implementation of work permit quota systems organized both internal and external migration flows in the late 1990s into a political-economic system to which migratory networks responded to. Thus, I argue that the 1991 census came at a critical point in Spanish immigration history. There were enough foreigners living in Spain at the time that counts by province and country of origin group were non-zero allowing temporal correlations to be drawn between the co-ethnic density measures, but the counts were very few as migration chains were hardly established at that point. Any network that would have been established would have been small, and the temporal distance between them and networks of immigrants arriving in the 2000s ensures independence of our instrument from the endogenous explanatory variable.^[We see in Figure 4 of the Appendix, less of those immigrants who arrived to Spain in 1990 had been influenced by co-ethnic contacts who had previous emigrated to Spain. It isn't until 1997 that we see for the first time a majority of the yearly arrival cohort having migrated under network influence, this influence increasing substantially in later cohorts.] 

``` {r echo = FALSE, warning = FALSE, message = FALSE}
## IV REGRESSION
model_data_iv <- ivreg(log_hour_earn ~ log_eth_dens_prov_1 + log_eth_dens_prov_1_relative + education_eng_class + years_final + age_final + age_sq_final + sex +  sector2 + country + province_strata | education_eng_class + years_final + age_final + age_sq_final + sex +  sector2 + country + province_strata + log_eth_dens_prov_1991 + log_eth_dens_prov_1991_relative, data = this_pop_1, se_type = "stata")
#coeftest(model_data_iv, vcov = vcovHC, type = "HC1")
#summary(model_data_iv, vcov = sandwich, diagnostics = TRUE)

model_data_iv_s1_abs <- lm(log_eth_dens_prov_1 ~ log_eth_dens_prov_1991 + log_eth_dens_prov_1991_relative + education_eng_class + years_final + age_final + age_sq_final + sex  + country + province_strata, data = this_pop_1)
model_data_iv_s1_pred_abs <- model_data_iv_s1_abs$fitted.values
#coeftest(model_data_iv_s1_abs, vcov = vcovHC, type = "HC1")

model_data_iv_s1_rel <- lm(log_eth_dens_prov_1_relative ~ log_eth_dens_prov_1991 + log_eth_dens_prov_1991_relative + education_eng_class + years_final + age_final + age_sq_final + sex + country + province_strata, data = this_pop_1)
model_data_iv_s1_pred_rel <- model_data_iv_s1_rel$fitted.values
#coeftest(model_data_iv_s1_rel, vcov = vcovHC, type = "HC1")

model_data_iv_s2 <- selection(wage_data ~ sex + country_class + woman_married_kids,
                              log_hour_earn ~ model_data_iv_s1_pred_abs + model_data_iv_s1_pred_rel + education_eng_class + years_final + age_final + age_sq_final + sex +  sector2 + country,
                              method = "2step",
                              data = this_pop_1)
  
#coeftest(model_data_iv_s2, vcov = vcovHC, type = "HC1")
#summary(model_data_iv_s2, vcov = sandwich, diagnostics = T)


rename_coefs <- function(model, from_vars, to_vars) {
  if (length(from_vars) != length(to_vars)) {
    print("ABORT! Length of from_vars not equal to length of to_vars!")
  } 
else {
  for (v in 1:length(from_vars)){
      from_var = from_vars[v]
      to_var = to_vars[v]
      names(model$coefficients)[which(names(model$coefficients) == from_var)] <- to_var
    }
  return(model)
  }
}

model_data_iv_s1_abs <- rename_coefs(model_data_iv_s1_abs, 
                                     c("log_eth_dens_prov_1"), 
                                     c("y"))
model_data_iv_s1_rel <- rename_coefs(model_data_iv_s1_rel, 
                                     c("log_eth_dens_prov_1_relative"), 
                                     c("y"))
model_data_iv_s2 <- rename_coefs(model_data_iv_s2, 
                                 c("log_hour_earn", "model_data_iv_s1_pred_abs", "model_data_iv_s1_pred_rel"), 
                                 c("y", "log_eth_dens_prov_1", "log_eth_dens_prov_1_relative"))
model_data_relativeviii_adj <- rename_coefs(model_data_relativeviii_adj, c("log_hour_earn"), c("y"))

```

```{r tabivreg, echo = FALSE, warning = FALSE, message = FALSE, results='asis', resize.height=4,resize.width=11}
#### WAGE TABLE (ABSOLUTE and RELATIVE ETHNIC DENSITY) ####
  stargazer(model_data_relativeviii_adj, model_data_iv_s1_abs, model_data_iv_s1_rel, model_data_iv_s2,
          type="latex",
          title = "Effect of log of co-ethnic density (absolute and relative) on log of hourly wage, OLS and 2SLS",
          colnames = F,
          dep.var.labels = c("Log of hourly wage", "logCOED (absolute)", "logCOED (relative)", "Log of hourly wage"),
          dep.var.labels.include = T,
          dep.var.caption = c(""),
          model.names = F,
          covariate.labels = c("logCOED (absolute)", "logCOED (relative)","logCOED1991 (absolute)", "logCOED1991 (relative)", "Female", "Years in Spain" , "Secondary (1st cycle)", "Secondary (2nd cycle)", "Tertiary (1st cycle)", "Tertiary (2nd cycle)", "Age", "Age squared"),
          keep = c("log_eth_dens_prov_1", "log_eth_dens_prov_1_relative", "sex", "years_final", "education_eng_class", "age_final", "age_sq_final"),
          omit = c("sector2", "country", "province_strata"),
          add.lines=list(c('Sector controls', rep("Yes", 4)),
                         c('Origin country controls', rep("Yes", 4)),
                         c('Region controls', rep("Yes", 4))),
          column.sep.width = "1pt",
          column.labels   = c("OLS", "2SLS - 1st stage", "2SLS - 2nd stage"),
          column.separate = c(1, 2, 1),
          header=FALSE,
          no.space = TRUE,
          selection.equation = FALSE,
          digits = 3,
          df = FALSE,
          single.row = F,
          font.size = "scriptsize",
          notes = c("Results shown for wage estimations in (1) and (4) are based on the Heckman selection model."),
          notes.align = "l",
          keep.stat = c("n", "rsq", "f", "ser", "adj.rsq", "chi2"),
          table.placement = "H")

```

Table 6 presents the results of the 2SLS regression that instruments both co-ethnic density measures in province $p$ in the year in which the individual of reference began living there with the respective measures from the same province in 1991 (Model 1).^[Some countries in 1991 did not map perfectly to countries in 2007 due to the dissolution of the Soviet Union and various other political happenings. Country mappings can be seen in Table \ref{tab:countrygroup} of the \ref{appendix}.] The effects, which can be understood as consistent estimates of the parameters of interest so long as the assumption holds that the instruments only act on the response variable through the endogenous variables, show a similar effect for absolute co-ethnic density as the one estimated by OLS (Model 4). For relative co-ethnic density, we see an even greater positive effect, though both estimates are statistically insignificant at $\alpha = 0.1$. Still, the results are convincing and provide evidence that network strength, regardless of the other covariates, has positive effects on immigrant wages while network magnitude has the opposite.


# CONCLUSION {#conclusion}
This analysis has attempted to identify and estimate the independent effect of residential settlement patterns on the wages of economic immigrants in Spain. However, like any other social phenomenon, there are meso-level factors that need to be considered, but separating the individual from the whole can be a daunting task. In this case of my study, co-ethnic density has come to represent immigrant networks since the non-random decision to settle in one region over another is often driven by where other similar individuals have settled before. Though unforced migratory movements present a great issue of selection bias into certain networks, assessing critically the endogeneity concerns has actually proven to be a rather useful exercise in understanding the different dimensions of migratory networks. By controlling for both absolute and relative co-ethnic densities, I disentagle the negative effects of increased immigrant labor stock and the positive effects of actively leveraging migratory networks on immigrant wages. My findings point to the conclusion that networks are beneficial when they are strong, but not necessarily when they are big. 

```{r echo = FALSE, warning = FALSE, message = FALSE}
comp_table_num <- rbind(
  cbind(var = "Years in Spain", 
        mean_econ = round(mean(this_pop_economic$years, na.rm = T), 2),
        sd_econ = round(sd(this_pop_economic$years, na.rm = T), 2),
        n_econ = sum(!is.na(this_pop_economic$years)),
        mean_nonecon = round(mean(this_pop_non_economic$years, na.rm = T), 2),
        sd_nonecon = round(sd(this_pop_non_economic$years, na.rm = T), 2),
        n_nonecon = sum(!is.na(this_pop_non_economic$years)),
        p = signif(t.test(this_pop_economic$years, this_pop_non_economic$years, alternative = "two.sided", var.equal = F)[3]$p.value, 2)),
  cbind(var = "Age", 
        mean_econ = round(mean(this_pop_economic$age, na.rm = T), 2),
        sd_econ = round(sd(this_pop_economic$age, na.rm = T), 2),
        n_econ = sum(!is.na(this_pop_economic$age)),
        mean_nonecon = round(mean(this_pop_non_economic$age, na.rm = T), 2),
        sd_nonecon = round(sd(this_pop_non_economic$age, na.rm = T), 2),
        n_nonecon = sum(!is.na(this_pop_non_economic$age)),
        p = signif(t.test(this_pop_economic$age, this_pop_non_economic$age, alternative = "two.sided", var.equal = F)[3]$p.value, 2))
  ) %>% as.data.frame()

```

```{r echo = FALSE, warning = FALSE, message = FALSE}
country_groupings <- read.xlsx(paste0(data_dir, "/country_groupings.xlsx")) %>% filter((LITERAL %in% levels(factor(this_pop_1$country))) | LITERAL %in% c("Guinea Ecuatorial", "Reino Unido","República Dominicana"))
```

\newpage
\pagenumbering{gobble}
# Appendix {#appendix}

```{r tabmeans, echo = FALSE, warning = FALSE, message = FALSE, fig.pos = "H", fig.align = "center"}
kable(comp_table_num, format = "latex", caption = "Comparison of Distributions of Age and Tenure in Spain for Economic and Non-Economic Immigrants", col.names = c("", "Mean", "SD", "Count","Mean", "SD", "Count", "p-value")) %>% 
  kable_styling(latex_options = c("HOLD_position", "scale_down")) %>%
add_header_above(c("Variable" = 1, "Economic Immigrants" = 3, "Non-economic Immigrants" = 3, "Test of Equal Means" = 1))
```

```{r sampselection, echo = FALSE, warning = FALSE, message = FALSE, fig.align = "center", fig.pos = "H"}
kable(prop_test_sampselection, format = "latex", caption = "Comparison of Cohorts of Wage Earning and Non-Wage Earning Economic Immigrants", col.names = c("", "Count", "Total", "Percent", "Count", "Total", "Percent", "p-value")) %>% 
  kable_styling(latex_options = c("HOLD_position", "scale_down")) %>%
add_header_above(c("Variable" = 1, "Wage Earners" = 3, "Non-Wage Earners" = 3, "Test of Equal Props." = 1)) %>%
  pack_rows(index = c("Gender" = 2, "Level of studies" = 5, "Country classification" = 7, "Married woman with kids?" = 2))
```

```{r countrygroup, echo = FALSE, warning = FALSE, message = FALSE, fig.pos = "H"}
kable(country_groupings, "latex", caption = "Countries Represented in the ENI and Mappings to Countries in Padrón data and 1991 Census") %>%
  kable_styling(latex_options = c("solid", "scale_down")) %>%
  row_spec(0,bold=TRUE)
```

```{r echo = FALSE, warning = FALSE, message = FALSE}
## Population density attributes by municipality and country of origin
padron_mun_data_name <- "poblacion_municipio_serie_temporal_2003_2007.xlsx"

dat_padron_mun <- read.xlsx(paste0(data_dir, "/", padron_mun_data_name), sheet = 1, startRow = 7, cols = c(3:6, 7, 9, 27, 19, 37, 34)) %>%
  mutate(across(c("Total", "Extranjero", "Rumanía", "Marruecos", "Colombia", "Ecuador"), as.numeric)) %>%
  melt(id.vars = c("PROVINCE_CODE", "MUNCIPALITY_CODE", "YEAR", "SEX")) %>%
  dplyr::rename(Origin = variable, Count = value) %>%
  group_by(PROVINCE_CODE, YEAR, SEX, Origin) %>%
  mutate(percent_prov_pop = (Count/sum(Count, na.rm = T))*100) %>%
  group_by(YEAR, SEX, Origin) %>%
  mutate(percent_total_pop = (Count/sum(Count, na.rm = T))*100) 

# Data for plotting percent of total municipal population from specific country of origin 
dat_padron_mun_overall <- left_join(x = dat_padron_mun %>% filter(PROVINCE_CODE != "TOTAL"),
                                           y = dat_padron_mun %>% filter(PROVINCE_CODE != "TOTAL" & Origin == "Total") %>% dplyr::select(-c("Origin")), 
                                           by = c("YEAR", "PROVINCE_CODE", "MUNCIPALITY_CODE", "SEX"),
                                           suffix = c("", "_Total"),
                                           keep = F) %>%
  mutate(percent_mun = (Count/Count_Total)*100,
         MUV = ifelse(Count %in% c(0:10000), "E1",
                      ifelse(Count %in% c(10001:20000), "E2",
                             ifelse(Count %in% c(20001:50000), "E3",
                                    ifelse(Count > 50000, "E4", "E9")))),
         MUV_ind = ifelse(Count <= 50000, "<= 50,000", "> 50,000"))

## Get municipality shapes
municipalities <- esp_get_munic_siane()
## Merge attributes with geo data
dat_geo_mun <- sp::merge(x = municipalities,
                    y = dat_padron_mun_overall,
                    by.x = "LAU_CODE",
                    by.y = "MUNCIPALITY_CODE",
                    all.x = TRUE
)

dat_geo_mun$MUV_factor <- factor(dat_geo_mun$MUV, levels = c("E1", "E2", "E3", "E4"), labels = c("< 10,000", "10,001 - 20,000", "20,001 - 50,000", "> 50,000"))

population_map_mun_total <- tm_shape(dat_geo_mun %>% filter(YEAR == "2007" & (Origin == "Total") & SEX == "All")) +
tm_borders(col = "grey") +
tm_polygons("MUV_factor", style = "cat", palette = "Reds",
            title = "Population Count",
            id = "LAU_CODE",
            ) +
tm_credits(text = "Data source: INE's padron continuo", position = c("right", "bottom"))
```

```{r figmun, fig.pos = "H", fig.cap="Municipalities by Population Count (2007)", echo = FALSE, warning = FALSE, message = FALSE, results='hide', fig.height = 4, fig.width = 8, fig.align = "center"}
population_map_mun_total + tm_shape(provinces) + tm_borders(col = "black")

```

```{r echo = FALSE, warning = FALSE, message = FALSE}
prop_influenced <- table(this_data$influenced, this_data$arrival_spain_yr, this_data$economic_final) %>% as.data.frame() %>% group_by(Var2, Var3) %>% mutate(prop = (Freq/sum(Freq))*100*20) %>% dplyr::rename(influenced = Var1, arrival_spain_yr = Var2, economic_plot = Var3) %>% mutate(influenced = as.numeric(paste0(influenced)), arrival_spain_yr = as.numeric(paste0(arrival_spain_yr))) %>% dplyr::select(-Freq) %>% filter(influenced == 1) 
```

```{r influenced, echo = FALSE, warning = FALSE, message = FALSE, fig.pos = "H", resize.height=5, resize.width=8, fig.cap = "Counts and percentage of ENI respondents whose decision to migrate to Spain was influenced by a family member, friend, or neighbor from home country who had already emigrated to Spain, by arrival year (1954 - 2007)"}
ggplot(left_join(x = this_data %>% mutate(economic_plot = factor(economic_final)), y = prop_influenced, by = c("influenced", "arrival_spain_yr", "economic_plot"), aes(x = arrival_spain_yr))) + 
  geom_histogram(aes(x = arrival_spain_yr, fill = economic_final), alpha = 0.4, position = 'identity', bin.width = 1) + 
  geom_smooth(aes(x = arrival_spain_yr, y = prop, color = economic_final)) + 
  scale_y_continuous(name = "Count", sec.axis = sec_axis(~./20, name="% whose decision to immigrate was influenced")) +
  scale_color_discrete(guide = "none") +
  xlim(c(1954, 2007)) +
  labs(
       fill = "Economic immigrant?") +
  theme_bw() +
  theme(legend.position="bottom") +
  xlab("Arrival Year")
```



\newpage

# References {#references}
<div id="refs"></div>
