==> 3_model_pop_output.Rmd <==
---
title: "ENI Analysis: Preliminary regressions"
author: "Madeline Swarr"
date: "4/28/2022"
output: pdf_document
header-includes:
    - \usepackage{setspace}\doublespacing
    - \usepackage[bottom]{footmisc}
    - \usepackage{float}
classoption: landscape
---

```{r echo = FALSE, warning = FALSE, message = FALSE}
########### ESPECIFICAMOS EL DIRECTORIO DEL CODIGO ################
head.dir <- "/Users/papapeach/Documents/UC3M/Thesis/Data Analysis/ENI-Analysis"
run_date <- "06jul2022"

# Descargar libraries 
dependency.dir <- paste0(head.dir, "/dependencies")
source(paste0(dependency.dir, "/get_packages.R"))
get_packages(c("MASS", "foreign", "Hmisc", "reshape2", "dplyr", "stargazer", "hypr", "sjPlot", "sjmisc"))

####################    Asignacion de parametros    #######################
# Recogemos la ruta del script que se esta ejecutando
input.dir <- paste0(head.dir, "/output/", run_date, "/1_processed")
output.dir <- paste0(head.dir, "/output/", run_date, "/3_modelled")
if (!dir.exists(output.dir)) {
  dir.create(output.dir)
}

this_data <- readRDS(paste0(input.dir, "/datos_eni_07_processed.rds"))
this_data$spanish <- factor(this_data$spanish, levels = c("Necesita mejorar", "Suficiente", "Bien", "Nativo")) %>% as.factor()
this_data$sex <- factor(this_data$sex, levels = c("Male", "Female")) %>% as.factor()
this_data$education <- factor(this_data$education, levels = c("Alguno/no sabe", "Sin estudios", "Primaria (incompleto)", "Primaria (completo)", "Secudaria (primer ciclo)", "Secudaria (segundo ciclo)", "Terciaria (primer ciclo)", "Terciaria (segundo ciclo)")) %>% as.factor()
this_data$sector2 <- factor(this_data$sector2, levels = c("Agriculture", "Business, commerce", "Construction", "Education", "Energy", "Extractive industries", "Extraterritorial organizations", "Financial intermediation", "Fishing", "Health and social services", "Hospitality", "Household activities", "Manufacturing", "Other social, community, personal services", "Public administration, defense", "Real estate; rental, business services", "Transportation, communications")) %>% as.factor()
this_data$work_permit <- factor(this_data$work_permit, levels = c("Yes", "No"))

contrasts(this_data$sector2, 16) <- centered_contrasts(hypr(contrasts(this_data$sector2)))
```

## Estimated hourly earnings
```{r echo = FALSE, warning = FALSE, message = FALSE}
######### 1b. Effect of concentration of coethnics in same province on estimated hourly earnings
this_data_2 <- this_data %>% filter(!is.na(log_hour_earn) & !is.na(log_eth_dens_prov) & !is.infinite(log_eth_dens_prov) & part_time == 0)

model_data_2i <- lm(log_hour_earn ~ log_eth_dens_prov, data = this_data_2)

model_data_2ii <- lm(log_hour_earn ~ log_eth_dens_prov + years + education, data = this_data_2)

model_data_2iii <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq, data = this_data_2)

model_data_2iv <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex, data = this_data_2)

model_data_2v <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit, data = this_data_2)

model_data_2vi <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2, data = this_data_2)

model_data_2vii <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short, data = this_data_2)

model_data_2viii <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province, data = this_data_2)

model_data_2ix <- lm(log_hour_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province + sector2*log_eth_dens_prov, data = this_data_2)

anova(model_data_2viii, model_data_2ix)

plot_model(model_data_2ix, type = "pred", terms = c("sector2", "log_eth_dens_prov")) + theme(axis.text.x = element_text(angle = 60))
```

```{r echo = FALSE, warning = FALSE, message = FALSE, results='asis', resize.height=4,resize.width=11}
stargazer(model_data_2i, model_data_2iv, model_data_2v, model_data_2vi, model_data_2vii, model_data_2viii,
          type="latex",
          title = "Effect of civic participation on log of estimated hourly earnings (for those who do not possess Spanish nationality)",
          colnames = FALSE,
          dep.var.labels = "Log of estimated hourly earnings",
          dep.var.labels.include = T,
          #covariate.labels = c("Civic engagement indicator", "Co-ethnic concentration", "Co-ethnic concentration squared","Count of siblings in same province", "Years since arrival", "Education", "Spanish - suficiente", "Spanish - bien", "Spanish - nativo", "Work authorization?", "Age", "Age squared", "Sex - female", "Civic engagement:Sex - female"),
          keep = c("log_eth_dens_prov", "years", "education", "work_permit", "age", "age_sq", "sex"),
          omit = c("sector2", "country_short", "province"),
          add.lines=list(c('Sector controls', rep("No", 3), rep("Yes", 3)),
                         c('Regional controls', rep("No", 4), rep("Yes", 2)),
                         c('Country of origin controls', rep("No", 5), rep("Yes", 1))),
          column.sep.width = "1pt",
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = T,
          font.size = "tiny",
          notes.align = "c",
          keep.stat = c("n", "rsq", "f", "ser"),
          table.placement = "H")

stargazer(model_data_2ix,
          type="latex",
          title = "Effect of civic participation on log of estimated hourly earnings (for those who do not possess Spanish nationality)",
          colnames = FALSE,
          dep.var.labels = "Log of estimated hourly earnings",
          dep.var.labels.include = T,
          #covariate.labels = c("Civic engagement indicator", "Co-ethnic concentration", "Co-ethnic concentration squared","Count of siblings in same province", "Years since arrival", "Education", "Spanish - suficiente", "Spanish - bien", "Spanish - nativo", "Work authorization?", "Age", "Age squared", "Sex - female", "Civic engagement:Sex - female"),
          keep = c("log_eth_dens_prov", "sector2:log_eth_dens_prov"),
          omit = c("years", "education", "work_permit", "age", "age_sq", "sex", "country_short", "province", "sector2"),
          add.lines=list(c('Years in Spain control', rep("Yes", 1)),
                         c('Education level control', rep("Yes", 1)),
                         c('Work authorization control', rep("Yes", 1)),
                         c('Age control', rep("Yes", 1)),
                         c('Age squared control', rep("Yes", 1)),
                         c('Gender controls', rep("Yes", 1)),
                         c('Country of origin controls', rep("Yes", 1)),
                         c('Regional controls', rep("Yes", 1)),
                         c('Sector controls', rep("Yes", 1))),
          column.sep.width = "1pt",
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = T,
          font.size = "tiny",
          notes.align = "c",
          keep.stat = c("n", "rsq", "f", "ser"),
          table.placement = "H")
```

## Log monthly earnings
```{r echo = FALSE, warning = FALSE, message = FALSE}
######### 3. Effect of concentration of coethnics in same province on log monthly income
this_data_3 <- this_data %>% filter(!is.na(log_month_earn) & !is.na(log_eth_dens_prov) & !is.infinite(log_eth_dens_prov) & part_time == 0)

model_data_3i <- lm(log_month_earn ~ log_eth_dens_prov, data = this_data_3)

model_data_3ii <- lm(log_month_earn ~ log_eth_dens_prov + years + education, data = this_data_3)

model_data_3iii <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq, data = this_data_3)

model_data_3iv <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex, data = this_data_3)

model_data_3v <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit, data = this_data_3)

model_data_3vi <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2, data = this_data_3)

model_data_3vii <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short, data = this_data_3)

model_data_3viii <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province, data = this_data_3)

model_data_3ix <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province + sector2:log_eth_dens_prov, data = this_data_3)

model_data_3x <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province + country_short:log_eth_dens_prov, data = this_data_3)

model_data_3xi <- lm(log_month_earn ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province + work_permit:log_eth_dens_prov, data = this_data_3)

# ANOVA for interaction of eth_dens_prov with sector and marginal effects plot
anova(model_data_3viii, model_data_3ix)
plot_model(model_data_3ix, type = "pred", terms = c("sector2", "log_eth_dens_prov")) + theme(axis.text.x = element_text(angle = 60))

# ANOVA for interaction of eth_dens_prov with country of origin and marginal effects plot
anova(model_data_3viii, model_data_3x)
plot_model(model_data_3x, type = "pred", terms = c("country_short [Alemania, Argentina, China, Colombia, Italia, Marruecos, Pakistán, Reino Unido, Ecuador, Venezuela, Rumanía, Senegal, Perú, Francia, Cuba]", "log_eth_dens_prov")) + theme(axis.text.x = element_text(angle = 60))

# ANOVA for interaction of eth_dens_prov with work authorization status and marginal effects plot
anova(model_data_3viii, model_data_3xi)
plot_model(model_data_3xi, type = "pred", terms = c("log_eth_dens_prov", "work_permit"))
```

```{r echo = FALSE, warning = FALSE, message = FALSE, results='asis', resize.height=4,resize.width=11}
stargazer(model_data_3i, model_data_3iv, model_data_3v, model_data_3vi, model_data_3vii, model_data_3viii,
          type="latex",
          title = "Effect of civic participation on log of estimated hourly earnings (for those who do not possess Spanish nationality)",
          colnames = FALSE,
          dep.var.labels = "Log of estimated hourly earnings",
          dep.var.labels.include = T,
          #covariate.labels = c("Civic engagement indicator", "Co-ethnic concentration", "Co-ethnic concentration squared","Count of siblings in same province", "Years since arrival", "Education", "Spanish - suficiente", "Spanish - bien", "Spanish - nativo", "Work authorization?", "Age", "Age squared", "Sex - female", "Civic engagement:Sex - female"),
          keep = c("log_eth_dens_prov", "years", "education", "work_permit", "age", "age_sq", "sex"),
          omit = c("sector2", "country_short", "province"),
          add.lines=list(c('Sector controls', rep("No", 3), rep("Yes", 3)),
                         c('Regional controls', rep("No", 4), rep("Yes", 2)),
                         c('Country of origin controls', rep("No", 5), rep("Yes", 1))),
          column.sep.width = "1pt",
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = T,
          font.size = "tiny",
          notes.align = "c",
          keep.stat = c("n", "rsq", "f", "ser"),
          table.placement = "H")

stargazer(model_data_3ix,
          type="latex",
          title = "Effect of civic participation on log of estimated hourly earnings (for those who do not possess Spanish nationality)",
          colnames = FALSE,
          dep.var.labels = "Log of estimated hourly earnings",
          dep.var.labels.include = T,
          #covariate.labels = c("Civic engagement indicator", "Co-ethnic concentration", "Co-ethnic concentration squared","Count of siblings in same province", "Years since arrival", "Education", "Spanish - suficiente", "Spanish - bien", "Spanish - nativo", "Work authorization?", "Age", "Age squared", "Sex - female", "Civic engagement:Sex - female"),
          keep = c("log_eth_dens_prov", "sector2:log_eth_dens_prov"),
          omit = c("years", "education", "work_permit", "age", "age_sq", "sex", "country_short", "province", "sector2"),
          add.lines=list(c('Years in Spain control', rep("Yes", 1)),
                         c('Education level control', rep("Yes", 1)),
                         c('Work authorization control', rep("Yes", 1)),
                         c('Age control', rep("Yes", 1)),
                         c('Age squared control', rep("Yes", 1)),
                         c('Gender controls', rep("Yes", 1)),
                         c('Country of origin controls', rep("Yes", 1)),
                         c('Regional controls', rep("Yes", 1)),
                         c('Sector controls', rep("Yes", 1))),
          column.sep.width = "1pt",
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = T,
          font.size = "tiny",
          notes.align = "c",
          keep.stat = c("n", "rsq", "f", "ser"),
          table.placement = "H")
```

## Job search
```{r echo = FALSE, warning = FALSE, message = FALSE}
######### 4. Effect of concentration of coethnics in same province on time it took to find a job
this_data_4 <- this_data %>% filter(!is.na(TPO1) & precontract == 0 & !is.na(log_eth_dens_prov) & !is.infinite(log_eth_dens_prov))

model_data_4i <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province, data = this_data_4, family = binomial())

model_data_4ii <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education , data = this_data_4, family = binomial())

model_data_4iii <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq, data = this_data_4, family = binomial())

model_data_4iv <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex, data = this_data_4, family = binomial())

model_data_4v <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit, data = this_data_4, family = binomial())

model_data_4vi <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2, data = this_data_4, family = binomial())

model_data_4vii <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short, data = this_data_4, family = binomial())

model_data_4iii <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province , data = this_data_4, family = binomial())

model_data_4ix <- glm(lt_search_1mo ~ log_eth_dens_prov + years + education + age + age_sq + sex + work_permit + sector2 + country_short + province + sector2:log_eth_dens_prov, data = this_data_4, family = binomial())

```

```{r echo = FALSE, warning = FALSE, message = FALSE, results='asis', resize.height=4,resize.width=11}
stargazer(model_data_4i, model_data_4iv, model_data_4v, model_data_4vi, model_data_4vii, model_data_4viii,
          type="latex",
          title = "Effect of log of co-ethnic density on probability of finding a job in less than a month",
          colnames = FALSE,
          dep.var.labels = "Probability of finding a job in less than a month",
          dep.var.labels.include = T,
          #covariate.labels = c("Civic engagement indicator", "Co-ethnic concentration", "Co-ethnic concentration squared","Count of siblings in same province", "Years since arrival", "Education", "Spanish - suficiente", "Spanish - bien", "Spanish - nativo", "Work authorization?", "Age", "Age squared", "Sex - female", "Civic engagement:Sex - female"),
          keep = c("log_eth_dens_prov", "years", "education", "work_permit", "age", "age_sq", "sex"),
          omit = c("sector2", "country_short", "province"),
          add.lines=list(c('Sector controls', rep("No", 3), rep("Yes", 3)),
                         c('Regional controls', rep("No", 4), rep("Yes", 2)),
                         c('Country of origin controls', rep("No", 5), rep("Yes", 1))),
          column.sep.width = "1pt",
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = T,
          font.size = "tiny",
          notes.align = "c",
          keep.stat = c("n", "rsq", "f", "ser"),
          table.placement = "H")

stargazer(model_data_4ix,
          type="latex",
          title = "Effect of log of co-ethnic density on probability of finding a job in less than a month",
          colnames = FALSE,
          dep.var.labels = "Probability of finding a job in less than a month",
          dep.var.labels.include = T,
          #covariate.labels = c("Civic engagement indicator", "Co-ethnic concentration", "Co-ethnic concentration squared","Count of siblings in same province", "Years since arrival", "Education", "Spanish - suficiente", "Spanish - bien", "Spanish - nativo", "Work authorization?", "Age", "Age squared", "Sex - female", "Civic engagement:Sex - female"),
          keep = c("log_eth_dens_prov", "sector2:log_eth_dens_prov"),
          omit = c("years", "education", "work_permit", "age", "age_sq", "sex", "country_short", "province", "sector2"),
          add.lines=list(c('Years in Spain control', rep("Yes", 1)),
                         c('Education level control', rep("Yes", 1)),
                         c('Work authorization control', rep("Yes", 1)),
                         c('Age control', rep("Yes", 1)),
                         c('Age squared control', rep("Yes", 1)),
                         c('Gender controls', rep("Yes", 1)),
                         c('Country of origin controls', rep("Yes", 1)),
                         c('Regional controls', rep("Yes", 1)),
                         c('Sector controls', rep("Yes", 1))),
          column.sep.width = "1pt",
          header=FALSE,
          no.space = TRUE,
          digits = 3,
          df = FALSE,
          single.row = T,
          font.size = "tiny",
          notes.align = "c",
          keep.stat = c("n", "rsq", "f", "ser"),
          table.placement = "H")
```
## General comments

### Variable fixes
1. Need to work on the **co-ethnic concentration measure**. Right now, it is measured as the percentage of the total registered population in each of the 52 Spanish provinces that are from a specific country of origin in 2006. As of now, we have municipal-level data by country and sex from 2003 onwards, but there is an issue with matching the municipalities from the padrón data to the ENI data that needs to be resolved first. Additionally, at what point in time should we be measuring co-ethnic concentration to represent social capital? Should it be an average of the years someone lived in a specific province/municipality? What if someone just recently moved to the province/municipality at the end of 2006? Any effect on wages wouldn't be due to this co-ethnic concentration. *Perhaps co-ethnic concentration effect is better used to predict time it takes to find a job or hours worked*.  

2. **Level of Spanish**, which is hard to measure, might be better to capture Spanish ability through years lived in Spain and country of origin. Also, strong ability to speak Spanish could also partially measure social capital -- one who is able to communicate clearly with another and create mutual understanding builds trust.

3. Control for previous work experience: using sector and occupation of work in the country of origin are not the best way. Maybe it could be an indicator variable of whether someone has previous work experience in the same sector in which they're working in currently (whether **SECT2** matches with **SECT1** (sector of work in home country) or with **SECT3** (sector of work in first job in Spain).

4. Investigate potential measurement bias with the response variable, **log of estimate hourly earnings**. Perhaps there is seasonality issues with wage data and hours worked as data was collected from November 2006 to February 2007. This could specifically affect those working in, say, education, seasonal farm work, the tourism industry. Perhaps a control for **sector in which one currently works (SECT2)** could help.

### Future research
1. Read more to see how I might construct a better measure of social capital to address structural dynamics/directionality. Would it be possible to do something like Sabatini did in Italy with PCA and SEM?
    * Investigate the **TERESP** and **CONVA** variables to see whether civic engagement indicator is actually measuring whether someone convalidated their titles/certificates in Spain. 
2. Look into other response variables -- social networks and unemployment and labor mobility?
    * For those who have a lot of ethnic social support upon arriving (but no pre-arranged job), how likely were they to find a first job quickly or be in PARO?
    * For those who got their first job through contacts, how likely were they to upgrade their jobs
3. Investigate bridging vs. bonding social capital
    * Look at the within-household structure and determine if immigrants are living with mainly family members, other co-ethnics, or Spanish nationals (we have birth country, nationality(ies), age, sex, year and month of arrival to Spain ) Could potentially filter to childless immigrants and determine if they have some personal connection to a Spanish national (whether via partners, friends, housemates).
